@@ -13,6 +13,7 @@ using System.Linq;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
@@ -43,7 +44,11 @@ namespace UndertaleModTool
                 new FrameworkPropertyMetadata(null));
 
         public static readonly PropertyInfo visualOffProp = typeof(Canvas).GetProperty("VisualOffset", BindingFlags.NonPublic | BindingFlags.Instance);
-        private static MainWindow mainWindow = Application.Current.MainWindow as MainWindow;
+        private static readonly MainWindow mainWindow = Application.Current.MainWindow as MainWindow;
+        private static readonly Regex trailingNumberRegex = new(@"\d+$", RegexOptions.Compiled);
+
+        // used for the flashing animation when a room object is selected
+        public static Dictionary<UndertaleObject, FrameworkElement> ObjElemDict { get; } = new();
 
         public UndertalePath PreviewPath
         {
@@ -52,7 +57,7 @@ namespace UndertaleModTool
         }
 
         private Stack<UndertaleObject> undoStack = new();
-        private CanvasWithDict roomCanvas;
+        private Canvas roomCanvas;
 
         public static readonly DoubleAnimation flashAnim = new(1, 0, TimeSpan.FromSeconds(0.75))
         {
@@ -70,14 +75,15 @@ namespace UndertaleModTool
             InitializeComponent();
 
             Loaded += UndertaleRoomEditor_Loaded;
+            Unloaded += UndertaleRoomEditor_Unloaded;
             DataContextChanged += UndertaleRoomEditor_DataContextChanged;
-        }
+        }        
 
         public void SaveImagePNG(Stream outfile)
         {
             if (roomCanvas is null)
             {
-                if (MainWindow.FindVisualChild<CanvasWithDict>(RoomGraphics) is CanvasWithDict canv && canv.Name == "RoomCanvas")
+                if (MainWindow.FindVisualChild<Canvas>(RoomGraphics) is Canvas canv && canv.Name == "RoomCanvas")
                     roomCanvas = canv;
                 else
                     throw new Exception("\"RoomCanvas\" not found.");
@@ -125,6 +131,10 @@ namespace UndertaleModTool
         {
             RoomRootItem.IsSelected = true;
         }
+        private void UndertaleRoomEditor_Unloaded(object sender, RoutedEventArgs e)
+        {
+            ObjElemDict.Clear();
+        }
 
         private void UndertaleRoomEditor_DataContextChanged(object sender, DependencyPropertyChangedEventArgs e)
         {
@@ -141,7 +151,7 @@ namespace UndertaleModTool
 
             UndertaleCachedImageLoader.Reset();
             CachedTileDataLoader.Reset();
-            roomCanvas?.ObjectDict.Clear();
+            ObjElemDict.Clear();
             roomObjDict.Clear();
             sprInstDict.Clear();
 
@@ -155,6 +165,22 @@ namespace UndertaleModTool
 
                 if (room.Layers.Count > 0) // if GMS 2+
                 {
+                    LayerZIndexConverter.ProcessOnce = true;
+
+                    bool ordered = true;
+                    for (int i = 0; i < room.Layers.Count - 1; i++)
+                    {
+                        if (room.Layers[i].LayerDepth > room.Layers[i + 1].LayerDepth)
+                        {
+                            ordered = false;
+                            break;
+                        }
+                    }
+
+                    if (!ordered)
+                        if (MainWindow.ShowQuestion("Room layers are not ordered by depth.\nRearrange them?", MessageBoxImage.Warning) == MessageBoxResult.Yes)
+                            room.RearrangeLayers();
+
                     Parallel.ForEach(room.Layers, (layer) =>
                     {
                         if (layer.LayerType == LayerType.Assets)
@@ -177,7 +203,7 @@ namespace UndertaleModTool
 
         private void RoomCanvas_Loaded(object sender, RoutedEventArgs e)
         {
-            roomCanvas = sender as CanvasWithDict;
+            roomCanvas = sender as Canvas;
         }
 
         private void TreeView_SelectedItemChanged(object sender, RoutedPropertyChangedEventArgs<object> e)
@@ -197,22 +223,23 @@ namespace UndertaleModTool
             else if (e.NewValue is UndertaleObject obj)
             {
                 ObjectEditor.Content = obj;
+
                 try
                 {
+                    if (obj is View)
+                        return;
+
                     DependencyObject obj1 = null;
+
                     if (obj is Layer layer)
                     {
-                        if (layer.LayerType == LayerType.Background)
-                            obj1 = VisualTreeHelper.GetChild(roomCanvas.ObjectDict[layer.BackgroundData], 0);
-                        else if (layer.LayerType == LayerType.Tiles)
-                            obj1 = VisualTreeHelper.GetChild(roomCanvas.ObjectDict[layer.TilesData], 0);
-                        else
+                        if (!layer.IsVisible)
                             return;
+
+                        obj1 = ObjElemDict[obj];
                     }
-                    else if (obj is View)
-                        return;
                     else
-                        obj1 = VisualTreeHelper.GetChild(roomCanvas.ObjectDict[obj], 0);
+                        obj1 = VisualTreeHelper.GetChild(ObjElemDict[obj], 0);
 
                     (obj1 as FrameworkElement).BringIntoView();
 
@@ -223,10 +250,24 @@ namespace UndertaleModTool
                     currStoryboard.Children.Add(flashAnim);
                     currStoryboard.Begin(this, true);
                 }
-                catch
+                catch (KeyNotFoundException)
                 {
                     Debug.WriteLine($"Flash animation error - \"{obj}\" is missing from the room object dictionary.");
                 }
+                catch (Exception ex)
+                {
+                    Debug.WriteLine($"Flash animation error - {ex}");
+                }
+            }
+            else if (e.NewValue is null && ObjectEditor.Content is Layer layer)
+            {
+                // if layers were rearranged and there was a layer selected
+                if ((DataContext as UndertaleRoom).Layers.Contains(layer))
+                {
+                    TreeViewItem layerItem = LayerItems.ItemContainerGenerator.ContainerFromItem(layer) as TreeViewItem;
+                    if (layerItem is not null)
+                        layerItem.IsSelected = true;
+                }    
             }
         }
 
@@ -261,124 +302,38 @@ namespace UndertaleModTool
             if (room.Layers.Count > 0)
                 layer = clickedObj switch
                 {
-                    Tile or GameObject => roomObjDict[(clickedObj as RoomObject).InstanceID],
+                    Tile or GameObject => roomObjDict[(clickedObj as IRoomObject).InstanceID],
                     SpriteInstance spr => sprInstDict[spr],
                     _ => null
                 };
 
             if (Keyboard.Modifiers.HasFlag(ModifierKeys.Alt))
             {
-                if (clickedObj is GameObject)
+                IList collection = clickedObj switch
                 {
-                    if (layer != null && layer.InstancesData == null)
-                    {
-                        MainWindow.ShowError("Must be on an instances layer");
-                        return;
-                    }
-                    var other = clickedObj as GameObject;
-                    var newObj = new GameObject
-                    {
-                        X = other.X,
-                        Y = other.Y,
-                        ObjectDefinition = other.ObjectDefinition,
-                        InstanceID = mainWindow.Data.GeneralInfo.LastObj++,
-                        CreationCode = other.CreationCode,
-                        ScaleX = other.ScaleX,
-                        ScaleY = other.ScaleY,
-                        Color = other.Color,
-                        Rotation = other.Rotation,
-                        PreCreateCode = other.PreCreateCode
-                    };
-
-                    if (layer != null)
-                    {
-                        int index = layer.InstancesData.Instances.IndexOf(other);
-                        layer.InstancesData.Instances.Insert(index + 1, newObj);
-                        roomObjDict.TryAdd(newObj.InstanceID, layer);
-                    }
-                    else
-                    {
-                        int index = room.GameObjects.IndexOf(other);
-                        room.GameObjects.Insert(index + 1, newObj);
-                    }
-
-                    clickedObj = newObj;
-                }
-                else if (clickedObj is Tile)
+                    GameObject => layer is null ? room.GameObjects : layer.InstancesData.Instances,
+                    Tile => layer is null ? room.Tiles : layer.AssetsData.LegacyTiles,
+                    SpriteInstance => layer.AssetsData.Sprites,
+                    _ => null
+                };
+                if (collection is not null)
                 {
-                    if (layer != null && layer.AssetsData == null)
+                    int index = collection.IndexOf(clickedObj) + 1;
+                    Point pos = clickedObj switch
                     {
-                        MainWindow.ShowError("Must be on an assets layer");
-                        return;
-                    }
-                    var other = clickedObj as Tile;
-                    var newObj = new Tile
-                    {
-                        X = other.X,
-                        Y = other.Y,
-                        _SpriteMode = other._SpriteMode,
-                        ObjectDefinition = other.ObjectDefinition,
-                        SourceX = other.SourceX,
-                        SourceY = other.SourceY,
-                        Width = other.Width,
-                        Height = other.Height,
-                        TileDepth = other.TileDepth,
-                        InstanceID = mainWindow.Data.GeneralInfo.LastTile++,
-                        ScaleX = other.ScaleX,
-                        ScaleY = other.ScaleY,
-                        Color = other.Color
+                        IRoomObject roomObj => new(roomObj.X, roomObj.Y),
+                        SpriteInstance sprInst => new(sprInst.X, sprInst.Y),
+                        _ => new()
                     };
-
-                    if (layer != null)
-                    {
-                        int index = layer.AssetsData.LegacyTiles.IndexOf(other);
-                        layer.AssetsData.LegacyTiles.Insert(index + 1, newObj);
-                        roomObjDict.TryAdd(newObj.InstanceID, layer);
-                    }
-                    else
-                    {
-                        int index = room.Tiles.IndexOf(other);
-                        room.Tiles.Insert(index + 1, newObj);
-                    }
-
-                    clickedObj = newObj;
-                }
-                else if (clickedObj is SpriteInstance)
-                {
-                    if (layer != null && layer.AssetsData == null)
-                    {
-                        MainWindow.ShowError("Must be on an assets layer");
-                        return;
-                    }
-
-                    string graphicName = "graphic_" + ((uint)new Random().Next(-int.MaxValue, int.MaxValue)).ToString("X8");
-                    var other = clickedObj as SpriteInstance;
-                    var newObj = new SpriteInstance
-                    {
-                        Name = new UndertaleString(graphicName),
-                        Sprite = other.Sprite,
-                        X = other.X,
-                        Y = other.Y,
-                        ScaleX = other.ScaleX,
-                        ScaleY = other.ScaleY,
-                        Color = other.Color,
-                        AnimationSpeed = other.AnimationSpeed,
-                        AnimationSpeedType = other.AnimationSpeedType,
-                        FrameIndex = other.FrameIndex,
-                        Rotation = other.Rotation,
-                    };
-
-                    int index = layer.AssetsData.Sprites.IndexOf(other);
-                    layer.AssetsData.Sprites.Insert(index + 1, newObj);
-                    sprInstDict.TryAdd(newObj, layer);
-
-                    clickedObj = newObj;
+                    clickedObj = AddObjectCopy(room, layer, clickedObj, true, index, pos);
                 }
             }
+            if (clickedObj is null)
+                return;
 
             SelectObject(clickedObj);
 
-            var mousePos = e.GetPosition(RoomGraphics);
+            var mousePos = e.GetPosition(roomCanvas);
             if (clickedObj is GameObject || clickedObj is Tile || clickedObj is SpriteInstance)
             {
                 movingObj = clickedObj;
@@ -409,7 +364,7 @@ namespace UndertaleModTool
                     {
                         X = other.X,
                         Y = other.Y,
-                        _SpriteMode = other._SpriteMode,
+                        spriteMode = other.spriteMode,
                         ObjectDefinition = other.ObjectDefinition,
                         SourceX = other.SourceX,
                         SourceY = other.SourceY,
@@ -448,10 +403,137 @@ namespace UndertaleModTool
                 }
             }
         }
+        private void Rectangle_MouseUp(object sender, MouseButtonEventArgs e)
+        {
+            e.Handled = true;
+            movingObj = null;
+        }
 
         bool placingTiles = false;
         List<Point> placedTiles = new();
 
+        // if "insertIndex" equals -1, then append
+        private UndertaleObject AddObjectCopy(UndertaleRoom room, Layer layer, UndertaleObject obj, bool showErrors, int insertIndex = -1, Point pos = new())
+        {
+            if (room is null || obj is null)
+                return null;
+
+            UndertaleObject newObj = null;
+
+            if (obj is Tile tile)
+            {
+                if (layer != null && layer.AssetsData == null)
+                {
+                    MainWindow.ShowError("Please select an assets layer.");
+                    return null;
+                }
+
+                var newTile = new Tile
+                {
+                    X = (int)pos.X,
+                    Y = (int)pos.Y,
+                    spriteMode = tile.spriteMode,
+                    ObjectDefinition = tile.ObjectDefinition,
+                    SpriteDefinition = tile.SpriteDefinition,
+                    SourceX = tile.SourceX,
+                    SourceY = tile.SourceY,
+                    Width = tile.Width,
+                    Height = tile.Height,
+                    TileDepth = tile.TileDepth,
+                    InstanceID = mainWindow.Data.GeneralInfo.LastTile++,
+                    ScaleX = tile.ScaleX,
+                    ScaleY = tile.ScaleY,
+                    Color = tile.Color
+                };
+
+                newObj = newTile;
+
+                int index;
+                if (layer != null)
+                {
+                    index = insertIndex > -1 ? insertIndex : layer.AssetsData.LegacyTiles.Count;
+                    layer.AssetsData.LegacyTiles.Insert(index, newTile);
+                    roomObjDict.TryAdd(newTile.InstanceID, layer);
+                }
+                else
+                {
+                    index = insertIndex > -1 ? insertIndex : room.Tiles.Count;
+                    room.Tiles.Insert(index, newTile);
+                }
+
+                // recalculates room grid size
+                room.SetupRoom();
+            }
+            else if (obj is GameObject gameObj)
+            {
+                if (layer != null && layer.InstancesData == null)
+                {
+                    MainWindow.ShowError("Please select an instance layer.");
+                    return null;
+                }
+
+                var newGameObj = new GameObject
+                {
+                    X = (int)pos.X,
+                    Y = (int)pos.Y,
+                    ObjectDefinition = gameObj.ObjectDefinition,
+                    InstanceID = mainWindow.Data.GeneralInfo.LastObj++,
+                    CreationCode = gameObj.CreationCode,
+                    ScaleX = gameObj.ScaleX,
+                    ScaleY = gameObj.ScaleY,
+                    Color = gameObj.Color,
+                    Rotation = gameObj.Rotation,
+                    PreCreateCode = gameObj.PreCreateCode,
+                    ImageSpeed = gameObj.ImageSpeed,
+                    ImageIndex = gameObj.ImageIndex
+                };
+
+                newObj = newGameObj;
+
+                int index = insertIndex > -1 ? insertIndex : room.GameObjects.Count;
+                room.GameObjects.Insert(index, newGameObj);
+                if (layer != null)
+                {
+                    index = insertIndex > -1 ? insertIndex : layer.InstancesData.Instances.Count;
+                    layer.InstancesData.Instances.Insert(index, newGameObj);
+                    roomObjDict.TryAdd(newGameObj.InstanceID, layer);
+                }
+            }
+            else if (obj is SpriteInstance sprInst)
+            {
+                if (layer != null && layer.AssetsData == null)
+                {
+                    MainWindow.ShowError("Please select an assets layer.");
+                    return null;
+                }
+
+                var newSprInst = new SpriteInstance
+                {
+                    Name = SpriteInstance.GenerateRandomName(mainWindow.Data),
+                    Sprite = sprInst.Sprite,
+                    X = (int)pos.X,
+                    Y = (int)pos.Y,
+                    ScaleX = sprInst.ScaleX,
+                    ScaleY = sprInst.ScaleY,
+                    Color = sprInst.Color,
+                    AnimationSpeed = sprInst.AnimationSpeed,
+                    AnimationSpeedType = sprInst.AnimationSpeedType,
+                    FrameIndex = sprInst.FrameIndex,
+                    Rotation = sprInst.Rotation,
+                };
+
+                newObj = newSprInst;
+
+                if (layer != null)
+                {
+                    int index = insertIndex > -1 ? insertIndex : layer.AssetsData.Sprites.Count;
+                    layer.AssetsData.Sprites.Insert(index, newSprInst);
+                    sprInstDict.TryAdd(newSprInst, layer);
+                }
+            }
+
+            return newObj;
+        }
         private void PaintObjects(Point gridMouse, UndertaleObject other, UndertaleRoom room)
         {
             if ((Mouse.LeftButton != MouseButtonState.Pressed) || !(Keyboard.Modifiers.HasFlag(ModifierKeys.Alt)))
@@ -467,7 +549,7 @@ namespace UndertaleModTool
             if (room.Layers.Count > 0)
                 layer = other switch
                 {
-                    Tile or GameObject => roomObjDict[(other as RoomObject).InstanceID],
+                    Tile or GameObject => roomObjDict[(other as IRoomObject).InstanceID],
                     SpriteInstance spr => sprInstDict[spr],
                     _ => null
                 };
@@ -480,91 +562,10 @@ namespace UndertaleModTool
             placedTiles.Add(gridMouse);
             placingTiles = true;
 
-            if (other is Tile)
-            {
-                var otherTile = other as Tile;
-                var newObj = new Tile
-                {
-                    X = (int)gridMouse.X,
-                    Y = (int)gridMouse.Y,
-                    _SpriteMode = otherTile._SpriteMode,
-                    ObjectDefinition = otherTile.ObjectDefinition,
-                    SpriteDefinition = otherTile.SpriteDefinition,
-                    SourceX = otherTile.SourceX,
-                    SourceY = otherTile.SourceY,
-                    Width = otherTile.Width,
-                    Height = otherTile.Height,
-                    TileDepth = otherTile.TileDepth,
-                    InstanceID = mainWindow.Data.GeneralInfo.LastTile++,
-                    ScaleX = otherTile.ScaleX,
-                    ScaleY = otherTile.ScaleY,
-                    Color = otherTile.Color
-                };
+            UndertaleObject newObj = AddObjectCopy(room, layer, other, false, -1, gridMouse);
 
-                if (layer != null)
-                {
-                    layer.AssetsData.LegacyTiles.Add(newObj);
-                    roomObjDict.TryAdd(newObj.InstanceID, layer);
-                }
-                else
-                    room.Tiles.Add(newObj);
-
+            if (newObj is not null)
                 SelectObject(newObj);
-            }
-            else if (other is GameObject)
-            {
-                var otherObject = other as GameObject;
-                var newObj = new GameObject
-                {
-                    X = (int)gridMouse.X,
-                    Y = (int)gridMouse.Y,
-                    ObjectDefinition = otherObject.ObjectDefinition,
-                    InstanceID = mainWindow.Data.GeneralInfo.LastObj++,
-                    CreationCode = otherObject.CreationCode,
-                    ScaleX = otherObject.ScaleX,
-                    ScaleY = otherObject.ScaleY,
-                    Color = otherObject.Color,
-                    Rotation = otherObject.Rotation,
-                    PreCreateCode = otherObject.PreCreateCode
-                };
-
-                room.GameObjects.Add(newObj);
-                if (layer != null)
-                {
-                    layer.InstancesData.Instances.Add(newObj);
-                    roomObjDict.TryAdd(newObj.InstanceID, layer);
-                }
-
-                SelectObject(newObj);
-            }
-            else if (other is SpriteInstance)
-            {
-                string graphicName = "graphic_" + ((uint)new Random().Next(-int.MaxValue, int.MaxValue)).ToString("X8");
-
-                var otherObject = other as SpriteInstance;
-                var newObj = new SpriteInstance
-                {
-                    Name = new UndertaleString(graphicName),
-                    Sprite = otherObject.Sprite,
-                    X = (int)gridMouse.X,
-                    Y = (int)gridMouse.Y,
-                    ScaleX = otherObject.ScaleX,
-                    ScaleY = otherObject.ScaleY,
-                    Color = otherObject.Color,
-                    AnimationSpeed = otherObject.AnimationSpeed,
-                    AnimationSpeedType = otherObject.AnimationSpeedType,
-                    FrameIndex = otherObject.FrameIndex,
-                    Rotation = otherObject.Rotation,
-                };
-
-                if (layer != null)
-                {
-                    layer.AssetsData.Sprites.Add(newObj);
-                    sprInstDict.TryAdd(newObj, layer);
-                }
-
-                SelectObject(newObj);
-            }
         }
 
         private void RectangleBackground_MouseDown(object sender, MouseButtonEventArgs e)
@@ -583,6 +584,8 @@ namespace UndertaleModTool
         {
             placingTiles = false;
             placedTiles.Clear();
+
+            movingObj = null;
         }
 
         private void RectangleBackground_MouseMove(object sender, MouseEventArgs e)
@@ -606,7 +609,7 @@ namespace UndertaleModTool
             {
                 UndertaleRoom room = this.DataContext as UndertaleRoom;
 
-                var mousePos = e.GetPosition(RoomGraphics);
+                var mousePos = e.GetPosition(roomCanvas);
 
                 int tgtX = (int)(mousePos.X - hotpointX);
                 int tgtY = (int)(mousePos.Y - hotpointY);
@@ -745,12 +748,6 @@ namespace UndertaleModTool
             }
         }
 
-        private void Rectangle_MouseUp(object sender, MouseButtonEventArgs e)
-        {
-            e.Handled = true;
-            movingObj = null;
-        }
-
         private UndertaleObject selectedObject;
         private void SelectObject(UndertaleObject obj)
         {
@@ -871,7 +868,14 @@ namespace UndertaleModTool
         {
             UndertaleObject sourceItem = e.Data.GetData(e.Data.GetFormats()[^1]) as UndertaleObject; // TODO: make this more reliable
 
-            e.Effects = e.AllowedEffects.HasFlag(DragDropEffects.Link) && sourceItem != null && (sourceItem is UndertaleGameObject || sourceItem is UndertalePath) ? DragDropEffects.Link : DragDropEffects.None;
+            e.Effects = e.AllowedEffects.HasFlag(DragDropEffects.Link) && sourceItem != null
+                        && ((mainWindow.IsGMS2 == Visibility.Collapsed && sourceItem is UndertaleBackground)
+                            || sourceItem is UndertaleGameObject
+                            || sourceItem is UndertalePath
+                            || (mainWindow.IsGMS2 == Visibility.Visible && sourceItem is UndertaleSprite))
+                        ? DragDropEffects.Link
+                        : DragDropEffects.None;
+
             e.Handled = true;
         }
 
@@ -879,36 +883,63 @@ namespace UndertaleModTool
         {
             UndertaleObject sourceItem = e.Data.GetData(e.Data.GetFormats()[^1]) as UndertaleObject;
 
-            e.Effects = e.AllowedEffects.HasFlag(DragDropEffects.Link) && sourceItem != null && (sourceItem is UndertaleGameObject || sourceItem is UndertalePath) ? DragDropEffects.Link : DragDropEffects.None;
+            e.Effects = e.AllowedEffects.HasFlag(DragDropEffects.Link) && sourceItem != null
+                        && ((mainWindow.IsGMS2 == Visibility.Collapsed && sourceItem is UndertaleBackground)
+                            || sourceItem is UndertaleGameObject
+                            || sourceItem is UndertalePath
+                            || (mainWindow.IsGMS2 == Visibility.Visible && sourceItem is UndertaleSprite))
+                        ? DragDropEffects.Link
+                        : DragDropEffects.None;
+
             if (e.Effects == DragDropEffects.Link)
             {
-                if (sourceItem is UndertaleBackground)
+                UndertaleRoom room = DataContext as UndertaleRoom;
+                Layer layer = null;
+                if (room.Layers.Count > 0)
                 {
-
+                    object utObj = ObjectEditor.Content;
+                    layer = utObj switch
+                    {
+                        GameObject gameObj => roomObjDict[gameObj.InstanceID],
+                        SpriteInstance sprInst => sprInstDict[sprInst],
+                        Layer l => l,
+                        _ => null
+                    };
                 }
-                else if (sourceItem is UndertaleGameObject)
+
+                if (sourceItem is UndertaleBackground droppedBG)
                 {
-                    UndertaleGameObject droppedObject = sourceItem as UndertaleGameObject;
+                    Background firstBG = room.Backgrounds.FirstOrDefault(x => x.BackgroundDefinition is null);
+                    if (firstBG is not null)
+                    {
+                        firstBG.Enabled = true;
+                        firstBG.BackgroundDefinition = droppedBG;
+
+                        SelectObject(firstBG);
+                    }
+                    else
+                        MainWindow.ShowError("No empty room backgrounds.");
+                }
+                else if (sourceItem is UndertaleGameObject droppedObj)
+                {
                     var mousePos = e.GetPosition(roomCanvas);
 
-                    UndertaleRoom room = this.DataContext as UndertaleRoom;
-                    Layer layer = ObjectEditor.Content as Layer;
                     if (mainWindow.IsGMS2 == Visibility.Visible && layer == null)
                     {
-                        MainWindow.ShowError("Must have a layer selected");
+                        MainWindow.ShowError("Please select a layer.");
                         return;
                     }
                     if (layer != null && layer.InstancesData == null)
                     {
-                        MainWindow.ShowError("Must be on an instances layer");
+                        MainWindow.ShowError("Please select an instances layer.");
                         return;
                     }
 
-                    var obj = new GameObject()
+                    GameObject obj = new()
                     {
                         X = (int)mousePos.X,
                         Y = (int)mousePos.Y,
-                        ObjectDefinition = droppedObject,
+                        ObjectDefinition = droppedObj,
                         InstanceID = mainWindow.Data.GeneralInfo.LastObj++
                     };
                     room.GameObjects.Add(obj);
@@ -918,15 +949,41 @@ namespace UndertaleModTool
 
                     SelectObject(obj);
                 }
-
-                if (sourceItem is UndertalePath)
+                else if (sourceItem is UndertalePath)
                 {
                     PreviewPath = sourceItem as UndertalePath;
                 }
-            }
-            e.Handled = true;
+                else if (sourceItem is UndertaleSprite droppedSprite)
+                {
+                    var mousePos = e.GetPosition(roomCanvas);
 
-            (this.DataContext as UndertaleRoom)?.SetupRoom(false);
+                    if (mainWindow.IsGMS2 == Visibility.Visible && layer == null)
+                    {
+                        MainWindow.ShowError("Please select a layer.");
+                        return;
+                    }
+                    if (layer != null && layer.AssetsData == null)
+                    {
+                        MainWindow.ShowError("Please select an assets layer.");
+                        return;
+                    }
+
+                    SpriteInstance sprInst = new()
+                    {
+                        X = (int)mousePos.X,
+                        Y = (int)mousePos.Y,
+                        Name = SpriteInstance.GenerateRandomName(mainWindow.Data),
+                        Sprite = droppedSprite
+                    };
+
+                    sprInstDict.TryAdd(sprInst, layer);
+                    layer.AssetsData.Sprites.Add(sprInst);
+
+                    SelectObject(sprInst);
+                }
+            }
+
+            e.Handled = true;
         }
 
         private void RoomObjectsTree_KeyDown(object sender, KeyEventArgs e)
@@ -1004,12 +1061,7 @@ namespace UndertaleModTool
                     room.Layers.Remove(layer);
 
                     if (layer.LayerType == LayerType.Background)
-                    {
-                        (RoomGraphics.ItemsSource as CompositeCollection).Remove(layer.BackgroundData);
                         room.UpdateBGColorLayer();
-                    }
-                    else if (layer.LayerType == LayerType.Tiles)
-                        (RoomGraphics.ItemsSource as CompositeCollection).Remove(layer.TilesData);
 
                     ObjectEditor.Content = null;
                 }
@@ -1029,16 +1081,16 @@ namespace UndertaleModTool
                 if (room.Layers.Count > 0)
                     layer = selectedObj switch
                     {
-                        Tile or GameObject => roomObjDict[(selectedObj as RoomObject).InstanceID],
+                        Tile or GameObject => roomObjDict[(selectedObj as IRoomObject).InstanceID],
                         SpriteInstance spr => sprInstDict[spr],
                         _ => null
                     };
 
                 IList list = selectedObj switch
                 {
-                    Tile tile => layer is null ? room.Tiles : layer.AssetsData.LegacyTiles,
-                    GameObject gameObj => layer is null ? room.GameObjects : layer.InstancesData.Instances,
-                    SpriteInstance sprInst => layer.AssetsData.Sprites,
+                    Tile => layer is null ? room.Tiles : layer.AssetsData.LegacyTiles,
+                    GameObject => layer is null ? room.GameObjects : layer.InstancesData.Instances,
+                    SpriteInstance => layer.AssetsData.Sprites,
                     _ => null
                 };
                 if (list is null)
@@ -1058,13 +1110,13 @@ namespace UndertaleModTool
                     if (layer is not null)
                     {
                         // swap back objects in "ObjectDict"
-                        var rect = roomCanvas.ObjectDict[selectedObj];
-                        var rectPrev = roomCanvas.ObjectDict[prevIndexObj as UndertaleObject];
-                        roomCanvas.ObjectDict[selectedObj] = rectPrev;
-                        roomCanvas.ObjectDict[prevIndexObj as UndertaleObject] = rect;
+                        var rect = ObjElemDict[selectedObj];
+                        var rectPrev = ObjElemDict[prevIndexObj as UndertaleObject];
+                        ObjElemDict[selectedObj] = rectPrev;
+                        ObjElemDict[prevIndexObj as UndertaleObject] = rect;
                     }
                 }
-
+                
                 SelectObject(selectedObj);
             }
         }
@@ -1133,88 +1185,35 @@ namespace UndertaleModTool
                 Debug.WriteLine("Paste");
                 Debug.WriteLine(obj);
             }*/
-
+            
             if (copied != null)
             {
                 UndertaleRoom room = this.DataContext as UndertaleRoom;
-
                 Layer layer = null;
                 if (room.Layers.Count > 0)
-                    layer = copied switch
+                {
+                    object utObj = ObjectEditor.Content;
+                    layer = utObj switch
                     {
-                        Tile or GameObject => roomObjDict[(copied as RoomObject).InstanceID],
+                        Tile or GameObject => roomObjDict[(utObj as IRoomObject).InstanceID],
                         SpriteInstance spr => sprInstDict[spr],
+                        Layer l => l,
                         _ => null
                     };
+                }
 
                 if (mainWindow.IsGMS2 == Visibility.Visible && layer == null)
                 {
-                    MainWindow.ShowError("Must paste onto a layer");
+                    MainWindow.ShowError("Please select a layer.");
                     return;
                 }
 
-                if (copied is GameObject)
-                {
-                    if (layer != null && layer.InstancesData == null)
-                    {
-                        MainWindow.ShowError("Must be on an instances layer");
-                        return;
-                    }
-                    var other = copied as GameObject;
-                    var obj = new GameObject();
-                    obj.X = other.X;
-                    obj.Y = other.Y;
-                    obj.ObjectDefinition = other.ObjectDefinition;
-                    obj.InstanceID = mainWindow.Data.GeneralInfo.LastObj++;
-                    obj.CreationCode = other.CreationCode;
-                    obj.ScaleX = other.ScaleX;
-                    obj.ScaleY = other.ScaleY;
-                    obj.Color = other.Color;
-                    obj.Rotation = other.Rotation;
-                    obj.PreCreateCode = other.PreCreateCode;
-                    room.GameObjects.Add(obj);
-                    if (layer != null)
-                    {
-                        roomObjDict.TryAdd(obj.InstanceID, layer);
-                        layer.InstancesData.Instances.Add(obj);
-                    }
+                Point mousePos = roomCanvas.IsMouseOver ? Mouse.GetPosition(roomCanvas) : new();
+                UndertaleObject newObj = AddObjectCopy(room, layer, copied, true, -1, mousePos);
 
-                    SelectObject(obj);
-                }
-                if (copied is Tile)
-                {
-                    if (layer != null && layer.AssetsData == null)
-                    {
-                        MainWindow.ShowError("Must be on an assets layer");
-                        return;
-                    }
-                    var other = copied as Tile;
-                    var obj = new Tile();
-                    obj.X = other.X;
-                    obj.Y = other.Y;
-                    obj._SpriteMode = other._SpriteMode;
-                    obj.ObjectDefinition = other.ObjectDefinition;
-                    obj.SourceX = other.SourceX;
-                    obj.SourceY = other.SourceY;
-                    obj.Width = other.Width;
-                    obj.Height = other.Height;
-                    obj.TileDepth = other.TileDepth;
-                    obj.InstanceID = mainWindow.Data.GeneralInfo.LastTile++;
-                    obj.ScaleX = other.ScaleX;
-                    obj.ScaleY = other.ScaleY;
-                    obj.Color = other.Color;
-                    if (layer != null)
-                    {
-                        roomObjDict.TryAdd(obj.InstanceID, layer);
-                        layer.AssetsData.LegacyTiles.Add(obj);
-                    }
-                    else
-                        room.Tiles.Add(obj);
-                    SelectObject(obj);
-                }
+                if (newObj is not null)
+                    SelectObject(newObj);
             }
-
-            (this.DataContext as UndertaleRoom)?.SetupRoom(false);
         }
 
         private void AddLayer<T>(LayerType type, string name) where T : Layer.LayerData, new()
@@ -1241,22 +1240,74 @@ namespace UndertaleModTool
                 }
             }
 
+            int layerDepth = 0;
+            if (room.Layers.Count > 0)
+            {
+                layerDepth = room.Layers.Select(l => l.LayerDepth).Max();
+                if (layerDepth + 100 > int.MaxValue)
+                {
+                    if (layerDepth + 1 > int.MaxValue)
+                    {
+                        layerDepth -= 1;
+                        MainWindow.ShowWarning("Warning - the maximum layer depth is reached.\nYou probably should change the depth of the new layer.");
+                    }
+                    else
+                        layerDepth += 1;
+                }
+                else
+                    layerDepth += 100;
+            }
+
+            string baseName = null;
+            int nameNum = 0;
+            while (room.Layers.Any(l => l.LayerName.Content == name))
+            {
+                if (baseName is null)
+                {
+                    // Get the trailing number from the name ("name123" => "123")
+                    Match numMatch = trailingNumberRegex.Match(name);
+
+                    // Name has a trailing number, so we parse the basename and number, increment the number and
+                    // set the new name to the basename and incremented number ("name123" -> "name124")
+                    if (numMatch.Success)
+                    {
+                        baseName = name[..^numMatch.Length];
+                        nameNum = Int32.Parse(numMatch.Groups[0].Value) + 1;
+
+                        name = baseName + nameNum;
+                    }
+                    // Name doesn't have a trailing number, so it's the first duplicate.
+                    // Thus we append append "1" to it ("name" -> "name1")
+                    else
+                        name += "1";
+                }
+                // If base name is already extracted, increment "nameNum" and append it to the base name
+                else
+                {
+                    nameNum++;
+                    name = baseName + nameNum;
+                }
+            }
+
             Layer layer = new();
             layer.LayerName = data.Strings.MakeString(name);
             layer.LayerId = largest_layerid + 1;
             layer.LayerType = type;
+            layer.LayerDepth = layerDepth;
             layer.Data = new T();
             room.Layers.Add(layer);
             room.UpdateBGColorLayer();
 
             if (layer.LayerType == LayerType.Assets)
             {
-                // create a new pointer list (if null)
+                // "??=" - assign if null
                 layer.AssetsData.LegacyTiles ??= new UndertalePointerList<Tile>();
-                // create new sprite pointer list (if null)
                 layer.AssetsData.Sprites ??= new UndertalePointerList<SpriteInstance>();
-                // create new sequence pointer list (if null)
                 layer.AssetsData.Sequences ??= new UndertalePointerList<SequenceInstance>();
+
+                // if it's needed to set "NineSlices"
+                if (!data.GMS2_3_2)
+                    layer.AssetsData.NineSlices ??= new UndertalePointerList<SpriteInstance>();
             }
             else if (layer.LayerType == LayerType.Tiles)
             {
@@ -1270,10 +1321,10 @@ namespace UndertaleModTool
 
         private void AddObjectInstance(UndertaleRoom room)
         {
-            var newobject = new GameObject { InstanceID = mainWindow.Data.GeneralInfo.LastObj++ };
-            room.GameObjects.Add(newobject);
+            var newObject = new GameObject { InstanceID = mainWindow.Data.GeneralInfo.LastObj++ };
+            room.GameObjects.Add(newObject);
 
-            SelectObject(newobject);
+            SelectObject(newObject);
         }
 
         private void AddGMS2ObjectInstance(Layer layer)
@@ -1303,7 +1354,7 @@ namespace UndertaleModTool
 
                 // add tile to list
                 var tile = new Tile { InstanceID = mainWindow.Data.GeneralInfo.LastTile++ };
-                tile._SpriteMode = true;
+                tile.spriteMode = true;
                 layer.AssetsData.LegacyTiles.Add(tile);
                 roomObjDict.TryAdd(tile.InstanceID, layer);
 
@@ -1318,12 +1369,15 @@ namespace UndertaleModTool
                 // add pointer list if one doesn't already exist
                 layer.AssetsData.Sprites ??= new UndertalePointerList<SpriteInstance>();
 
-                // add tile to list
-                var spriteinstance = new SpriteInstance();
-                layer.AssetsData.Sprites.Add(spriteinstance);
-                sprInstDict.TryAdd(spriteinstance, layer);
+                SpriteInstance spriteInstance = new()
+                {
+                    Name = SpriteInstance.GenerateRandomName(mainWindow.Data)
+                };
 
-                SelectObject(spriteinstance);
+                layer.AssetsData.Sprites.Add(spriteInstance);
+                sprInstDict.TryAdd(spriteInstance, layer);
+
+                SelectObject(spriteInstance);
             }
         }
 
@@ -1373,111 +1427,14 @@ namespace UndertaleModTool
             AddSprite(menuitem.DataContext as Layer);
         }
 
-        private void MenuItem_NewGMS1Tile_Click(object sender, RoutedEventArgs e)
-        {
-            AddGMS1Tile(this.DataContext as UndertaleRoom);
-        }
-
-        private void TileDataExport_Click(object sender, RoutedEventArgs e)
-        {
-            if (RoomObjectsTree.SelectedItem is Layer layer)
-            {
-                if (layer.TilesData.TileData.Length == 0)
-                {
-                    MainWindow.ShowError("Tile data is empty.");
-                    return;
-                }
-
-                StringBuilder sb = new();
-                foreach (uint[] dataRow in layer.TilesData.TileData)
-                    sb.AppendLine(String.Join(";", dataRow.Select(x => x.ToString())));
-
-                string dataFolder = System.IO.Path.GetDirectoryName((Application.Current.MainWindow as MainWindow).FilePath);
-                string filePath = System.IO.Path.Combine(dataFolder, $"{layer.LayerName.Content}_tiledata.csv");
-
-                try
-                {
-                    File.WriteAllText(filePath, sb.ToString());
-                }
-                catch (Exception ex)
-                {
-                    MainWindow.ShowError($"Error while saving the file - \"{ex.Message}\".");
-                    return;
-                }
-
-                MainWindow.ShowMessage("Exported file path: " + filePath);
-            }
-        }
-        private void TileDataImport_Click(object sender, RoutedEventArgs e)
-        {
-            if (RoomObjectsTree.SelectedItem is Layer layer)
-            {
-                if (layer.TilesData.TilesX == 0 || layer.TilesData.TilesY == 0)
-                {
-                    MainWindow.ShowError("Tile data size can't be zero.");
-                    return;
-                }
-
-                OpenFileDialog dlg = new()
-                {
-                    DefaultExt = "csv",
-                    Filter = "CSV table|*.csv|All files|*"
-                };
-
-                if (dlg.ShowDialog() == true)
-                {
-                    uint[][] tileDataNew = (uint[][])layer.TilesData.TileData.Clone();
-                    string[] tileDataLines = null;
-                    try
-                    {
-                        tileDataLines = File.ReadAllLines(dlg.FileName);
-                    }
-                    catch (Exception ex)
-                    {
-                        MainWindow.ShowError($"Error while opening file - \"{ex.Message}\".");
-                        return;
-                    }
-
-                    if (tileDataLines?.Length != tileDataNew.Length)
-                    {
-                        MainWindow.ShowError("Error - selected file line count doesn't match tile layer height.");
-                        return;
-                    }
-
-                    for (int i = 0; i < tileDataLines.Length; i++)
-                    {
-                        uint[] dataRow;
-                        try
-                        {
-                            dataRow = tileDataLines[i].Split(';').Select(x => UInt32.Parse(x)).ToArray();
-                        }
-                        catch (Exception ex)
-                        {
-                            MainWindow.ShowError($"Error while parsing line {i + 1} - \"{ex.Message}\".");
-                            return;
-                        }
-
-                        if (dataRow.Length != layer.TilesData.TilesX)
-                        {
-                            MainWindow.ShowError($"Length of line {i + 1} is not equal to the tile data width.");
-                            return;
-                        }
-
-                        tileDataNew[i] = dataRow;
-                    }
-
-                    layer.TilesData.TileData = tileDataNew;
-
-                    MainWindow.ShowMessage("Imported successfully.");
-                }
-            }
-        }
-
         private void MenuItem_NewObjectInstance_Click(object sender, RoutedEventArgs e)
         {
             AddObjectInstance(this.DataContext as UndertaleRoom);
         }
-
+        private void MenuItem_NewGMS1Tile_Click(object sender, RoutedEventArgs e)
+        {
+            AddGMS1Tile(this.DataContext as UndertaleRoom);
+        }
 
         public static void GenerateSpriteCache(UndertaleRoom room)
         {
@@ -1502,11 +1459,12 @@ namespace UndertaleModTool
                             else
                                 tiles.AddRange(layer.AssetsData.LegacyTiles);
 
-                            allObjects.AddRange(layer.AssetsData.Sprites);
+                            allObjects.AddRange(layer.AssetsData.Sprites.Where(s => s.Sprite?.Textures.Count > 0));
                             break;
 
                         case LayerType.Background:
-                            allObjects.Add(layer.BackgroundData);
+                            if (layer.BackgroundData.Sprite?.Textures.Count > 0)
+                                allObjects.Add(layer.BackgroundData);
                             break;
 
                         case LayerType.Instances:
@@ -1520,13 +1478,13 @@ namespace UndertaleModTool
                 tiles = room.Tiles.ToList();
 
                 allObjects.AddRange(room.Backgrounds);
-                allObjects.AddRange(room.GameObjects);
+                allObjects.AddRange(room.GameObjects.Where(g => g.ObjectDefinition?.Sprite?.Textures.Count > 0));
             }
 
             tileTextures = tiles?.AsParallel()
                                  .WithExecutionMode(ParallelExecutionMode.ForceParallelism)
                                  .GroupBy(x => x.Tpag?.Name?.Content ?? (mainWindow.Data.TexturePageItems.IndexOf(x.Tpag) + 1).ToString())
-                                 .Where(x => x.Key != "-1")
+                                 .Where(x => x.Key != "0")
                                  .Select(x =>
                                  {
                                      return new Tuple<UndertaleTexturePageItem, List<Tuple<uint, uint, uint, uint>>>(
@@ -1542,11 +1500,11 @@ namespace UndertaleModTool
                 UndertaleTexturePageItem texture = obj switch
                 {
                     Background bg => bg.BackgroundDefinition?.Texture,
-                    GameObject gameObj => gameObj.ObjectDefinition?.Sprite?.Textures[0].Texture,
+                    GameObject gameObj => gameObj.ObjectDefinition?.Sprite?.Textures[gameObj.WrappedImageIndex].Texture,
 
                     // GMS 2+
                     Layer.LayerBackgroundData bgData => bgData.Sprite?.Textures[0].Texture,
-                    SpriteInstance sprite => sprite.Sprite?.Textures[0].Texture,
+                    SpriteInstance sprite => sprite.Sprite?.Textures[sprite.WrappedFrameIndex].Texture,
                     _ => null
                 };
 
@@ -1556,7 +1514,7 @@ namespace UndertaleModTool
                     if (textPageName is null)
                     {
                         textPageName = (mainWindow.Data.TexturePageItems.IndexOf(texture) + 1).ToString();
-                        if (textPageName == "-1")
+                        if (textPageName == "0")
                             return;
                     }
 
@@ -1588,78 +1546,164 @@ namespace UndertaleModTool
                 loader.Convert(texture, null, "generate", null);
             });
         }
+
+        private void TileDataExport_Click(object sender, RoutedEventArgs e)
+        {
+            if (RoomObjectsTree.SelectedItem is not Layer layer)
+                return;
+
+            if (layer.TilesData.TileData.Length == 0)
+            {
+                MainWindow.ShowError("Tile data is empty.");
+                return;
+            }
+
+            StringBuilder sb = new();
+            foreach (uint[] dataRow in layer.TilesData.TileData)
+                sb.AppendLine(String.Join(";", dataRow.Select(x => x.ToString())));
+
+            string dataFolder = System.IO.Path.GetDirectoryName((Application.Current.MainWindow as MainWindow).FilePath);
+            string filePath = System.IO.Path.Combine(dataFolder, $"{layer.LayerName.Content}_tiledata.csv");
+
+            try
+            {
+                File.WriteAllText(filePath, sb.ToString());
+            }
+            catch (Exception ex)
+            {
+                MainWindow.ShowError($"Error while saving the file - \"{ex.Message}\".");
+                return;
+            }
+
+            MainWindow.ShowMessage("Exported file path: " + filePath);
+        }
+        private void TileDataImport_Click(object sender, RoutedEventArgs e)
+        {
+            if (RoomObjectsTree.SelectedItem is not Layer layer)
+                return;
+
+            if (layer.TilesData.TilesX == 0 || layer.TilesData.TilesY == 0)
+            {
+                MainWindow.ShowError("Tile data size can't be zero.");
+                return;
+            }
+
+            OpenFileDialog dlg = new()
+            {
+                DefaultExt = "csv",
+                Filter = "CSV table|*.csv|All files|*"
+            };
+
+            if (dlg.ShowDialog() == true)
+            {
+                uint[][] tileDataNew = (uint[][])layer.TilesData.TileData.Clone();
+                string[] tileDataLines = null;
+                try
+                {
+                    tileDataLines = File.ReadAllLines(dlg.FileName);
+                }
+                catch (Exception ex)
+                {
+                    MainWindow.ShowError($"Error while opening file - \"{ex.Message}\".");
+                    return;
+                }
+
+                if (tileDataLines?.Length != tileDataNew.Length)
+                {
+                    MainWindow.ShowError("Error - selected file line count doesn't match tile layer height.");
+                    return;
+                }
+
+                for (int i = 0; i < tileDataLines.Length; i++)
+                {
+                    uint[] dataRow;
+                    try
+                    {
+                        dataRow = tileDataLines[i].Split(';').Select(x => UInt32.Parse(x)).ToArray();
+                    }
+                    catch (Exception ex)
+                    {
+                        MainWindow.ShowError($"Error while parsing line {i + 1} - \"{ex.Message}\".");
+                        return;
+                    }
+
+                    if (dataRow.Length != layer.TilesData.TilesX)
+                    {
+                        MainWindow.ShowError($"Length of line {i + 1} is not equal to the tile data width.");
+                        return;
+                    }
+
+                    tileDataNew[i] = dataRow;
+                }
+
+                layer.TilesData.TileData = tileDataNew;
+
+                MainWindow.ShowMessage("Imported successfully.");
+            }
+        }
+
+        private void LayerCanvas_DataContextChanged(object sender, DependencyPropertyChangedEventArgs e)
+        {
+            // raised on layers rearrange
+            if (e.NewValue is not Layer layer)
+                return;
+
+            LayerCanvas canvas = sender as LayerCanvas;
+
+            canvas.CurrentLayer = layer;
+            ObjElemDict[layer] = canvas;
+        }
+        private void LayerCanvas_Initialized(object sender, EventArgs e)
+        {
+            LayerCanvas canvas = sender as LayerCanvas;
+
+            if (canvas?.DataContext is not Layer layer)
+                return;
+
+            canvas.CurrentLayer = layer;
+            ObjElemDict[layer] = canvas;
+        }
+        private void LayerCanvas_Unloaded(object sender, RoutedEventArgs e)
+        {
+            LayerCanvas canvas = sender as LayerCanvas;
+
+            if (canvas.CurrentLayer is not null)
+                ObjElemDict.Remove(canvas.CurrentLayer);
+        }
     }
 
-
-    public partial class CanvasWithDict : Canvas
+    public partial class RoomCanvas : Canvas
     {
-        public Dictionary<UndertaleObject, FrameworkElement> ObjectDict { get; } = new();
+        private readonly bool isGMS2 = (Application.Current.MainWindow as MainWindow).IsGMS2 == Visibility.Visible;
+
+        protected override void OnVisualChildrenChanged(DependencyObject visualAdded, DependencyObject visualRemoved)
+        {
+            if (!isGMS2)
+            {
+                if (visualAdded is ContentPresenter presAdded && presAdded.DataContext is UndertaleObject objAdded)
+                    UndertaleRoomEditor.ObjElemDict[objAdded] = presAdded;
+
+                if (visualRemoved is ContentPresenter presRemoved && presRemoved.DataContext is UndertaleObject objRemoved)
+                    UndertaleRoomEditor.ObjElemDict.Remove(objRemoved);
+            }
+
+            base.OnVisualChildrenChanged(visualAdded, visualRemoved);
+        }
+    }
+    public partial class LayerCanvas : Canvas
+    {
+        // "DataContext" on "Unloaded" event is "{DisconnectedItem}"
+        public UndertaleObject CurrentLayer { get; set; }
 
         protected override void OnVisualChildrenChanged(DependencyObject visualAdded, DependencyObject visualRemoved)
         {
             if (visualAdded is ContentPresenter presAdded && presAdded.DataContext is UndertaleObject objAdded)
-                ObjectDict[objAdded] = visualAdded as FrameworkElement;
+                UndertaleRoomEditor.ObjElemDict[objAdded] = presAdded;
 
             if (visualRemoved is ContentPresenter presRemoved && presRemoved.DataContext is UndertaleObject objRemoved)
-                ObjectDict.Remove(objRemoved);
-        }
-    }
+                UndertaleRoomEditor.ObjElemDict.Remove(objRemoved);
 
-
-    [ValueConversion(typeof(ObservableCollection<GameObject>), typeof(int))]
-    public class ObjCenterXConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            GameObject obj = value as GameObject;
-            if (obj == null)
-                return 0;
-            if (obj.ObjectDefinition == null || obj.ObjectDefinition.Sprite == null)
-                return obj.X;
-            return (obj.X + (obj.ObjectDefinition.Sprite.OriginX * obj.ScaleX));
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    [ValueConversion(typeof(ObservableCollection<GameObject>), typeof(int))]
-    public class ObjXOffsetConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            GameObject obj = value as GameObject;
-            if (obj == null)
-                return 0;
-            if (obj.ObjectDefinition == null || obj.ObjectDefinition.Sprite == null)
-                return obj.X;
-            return obj.X + obj.ObjectDefinition.Sprite.OriginX;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    [ValueConversion(typeof(ObservableCollection<GameObject>), typeof(int))]
-    public class ObjCenterYConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            GameObject obj = value as GameObject;
-            if (obj == null)
-                return 0;
-            if (obj.ObjectDefinition == null || obj.ObjectDefinition.Sprite == null)
-                return obj.Y;
-            return (obj.Y + (obj.ObjectDefinition.Sprite.OriginY * obj.ScaleY));
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
+            base.OnVisualChildrenChanged(visualAdded, visualRemoved);
         }
     }
 
@@ -1697,7 +1741,7 @@ namespace UndertaleModTool
 
         public object Convert(object[] values, Type targetType, object parameter, CultureInfo culture)
         {
-            if (values.Any(x => x is null))
+            if (values.Any(x => x is null || x == DependencyProperty.UnsetValue))
                 return null;
 
             bool isGMS2 = ((RoomEntryFlags)values[1]).HasFlag(RoomEntryFlags.IsGMS2);
@@ -1735,40 +1779,92 @@ namespace UndertaleModTool
         }
     }
 
-    public class LayerFlattenerConverter : IValueConverter
+    public class LayerItemsSourceConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            CompositeCollection collection = new();
-            IList<Layer> layers = value as IList<Layer>;
-            foreach (var layer in layers.OrderByDescending((x) => x?.LayerDepth ?? 0))
+            if (value is Layer layer)
             {
-                if (layer == null)
-                    continue;
-
-                switch (layer.LayerType)
+                return layer.LayerType switch
                 {
-                    case LayerType.Background:
-                        collection.Add(layer.BackgroundData);
-                        break;
-                    case LayerType.Instances:
-                        collection.Add(new CollectionContainer() { Collection = layer.InstancesData.Instances });
-                        break;
-                    case LayerType.Assets:
-                        collection.Add(new CollectionContainer() { Collection = layer.AssetsData.LegacyTiles });
-                        collection.Add(new CollectionContainer() { Collection = layer.AssetsData.Sprites });
-                        break;
-                    case LayerType.Tiles:
-                        collection.Add(layer.TilesData);
-                        break;
-                }
+                    // TODO: implement "LayerType.Effects"
+                    LayerType.Assets => new CompositeCollection(2)
+                                        {
+                                          new CollectionContainer() { Collection = layer.AssetsData.LegacyTiles },
+                                          new CollectionContainer() { Collection = layer.AssetsData.Sprites }
+                                        },
+                    LayerType.Background => new List<Layer.LayerBackgroundData>() { layer.BackgroundData },
+                    LayerType.Instances => layer.InstancesData.Instances,
+                    LayerType.Tiles => new List<Layer.LayerTilesData>() { layer.TilesData },
+                    _ => null
+                };
             }
-            return collection;
+            else
+                return null;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            throw new NotSupportedException();
+            throw new NotImplementedException();
+        }
+    }
+    public class LayerZIndexConverter : IMultiValueConverter
+    {
+        public static bool ProcessOnce { get; set; }
+
+        private static bool suspended;
+        private static int remainingCount = -1;
+        public object Convert(object[] values, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (values[0] is Layer layer && layer.ParentRoom is UndertaleRoom room)
+            {
+                int layerIndex = room.Layers.Count - room.Layers.IndexOf(layer) - 1;
+
+                if (ProcessOnce)
+                {
+                    if (remainingCount == -1 && room.Layers.Count > 0)
+                        remainingCount = room.Layers.Count - 1;
+                    else if (remainingCount > 0)
+                    {
+                        remainingCount--;
+
+                        if (remainingCount == 0)
+                        {
+                            ProcessOnce = false;
+                            remainingCount = -1;
+                        }
+                    }
+                }
+                else if (!suspended)
+                {
+                    bool ordered = true;
+                    for (int i = 0; i < room.Layers.Count - 1; i++)
+                    {
+                        if (room.Layers[i].LayerDepth > room.Layers[i + 1].LayerDepth)
+                        {
+                            ordered = false;
+                            break;
+                        }
+                    }
+
+                    if (!ordered)
+                    {
+                        suspended = true;
+                        var roomEditor = MainWindow.FindVisualChild<UndertaleRoomEditor>((Application.Current.MainWindow as MainWindow).DataEditor);
+                        room.RearrangeLayers(roomEditor?.RoomObjectsTree.SelectedItem as Layer);
+                        suspended = false;
+                    }
+                }
+
+                return layerIndex;
+            }
+            else
+                return -1;
+        }
+
+        public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
         }
     }
 
@@ -1793,13 +1889,12 @@ namespace UndertaleModTool
                         {
                             try
                             {
-                                MessageBox.Show("Room flags of the GMS 2+ games must contain \"IsGMS2\" flag, else game would crash on that room.",
-                                            "UndertaleModTool", MessageBoxButton.OK, MessageBoxImage.Error);
+                                MainWindow.ShowError("Room flags of GMS 2+ games must contain the \"IsGMS2\" flag, otherwise the game will crash when loading that room.", false);
                             }
                             catch {}
                         }
 
-                        flags |= RoomEntryFlags.IsGMS2 | RoomEntryFlags.IsGMS2_3;
+                        flags |= RoomEntryFlags.IsGMS2;
                     }
 
                     return flags;
@@ -1859,10 +1954,10 @@ namespace UndertaleModTool
                 texture = bg.BackgroundDefinition.Texture;
 
                 bg.UpdateStretch();
-                offsetV = bg.Y * (1 / bg.CalcScaleY);
-                offsetH = bg.X * (1 / bg.CalcScaleX);
-                xOffset = bg.X;
-                yOffset = bg.Y;
+                offsetV = bg.YOffset * (1 / bg.CalcScaleY);
+                offsetH = bg.XOffset * (1 / bg.CalcScaleX);
+                xOffset = bg.XOffset;
+                yOffset = bg.YOffset;
                 roomWidth = System.Convert.ToDouble(bg.ParentRoom.Width);
                 roomHeight = System.Convert.ToDouble(bg.ParentRoom.Height);
             }
@@ -1942,7 +2037,8 @@ namespace UndertaleModTool
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return value is LayerType.Instances ? Visibility.Collapsed : Visibility.Visible;
+            bool includeBGLayer = parameter is string par && par == "Canvas";
+            return value is LayerType.Instances || (includeBGLayer && value is LayerType.Background) ? Visibility.Collapsed : Visibility.Visible;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
@@ -2004,19 +2100,6 @@ namespace UndertaleModTool
             }
 
             return GridLength.Auto;
-        }
-
-        public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class BGLayerVisConverter : IMultiValueConverter
-    {
-        public object Convert(object[] values, Type targetType, object parameter, CultureInfo culture)
-        {
-            return (bool)values[0] && (bool)values[1] ? Visibility.Visible : Visibility.Collapsed;
         }
 
         public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)
@@ -2205,6 +2288,22 @@ namespace UndertaleModTool
         }
 
         public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+
+    public class ParentGridHeightConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (value is double h)
+                return h - 22; // "TabController" has predefined height
+            else
+                return 0;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
             throw new NotImplementedException();
         }
