@@ -2,940 +2,847 @@
 using Microsoft.CodeAnalysis.Scripting;
 using Newtonsoft.Json;
 using System;
-using System.Diagnostics;
+using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
-using System.Runtime.InteropServices;
 using UndertaleModLib;
 using UndertaleModLib.Scripting;
 using UndertaleModLib.Util;
 using static UndertaleModLib.UndertaleReader;
-
 using System.CommandLine;
 using System.CommandLine.Invocation;
 using System.CommandLine.Builder;
 using System.CommandLine.Parsing;
+using System.Linq;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Linq;
-using System.Collections.Generic;
-using UndertaleModLib.Decompiler;
 using UndertaleModLib.Models;
 
+namespace UndertaleModCli;
+
 /// <summary>
-/// The supplied location of the data file did not exist
+/// Main CLI Program
 /// </summary>
-public class DataFileNotFoundException : ArgumentException
+public partial class Program : IScriptInterface
 {
-    public DataFileNotFoundException() { }
-    public DataFileNotFoundException(string message) : base(message) { }
-    public DataFileNotFoundException(string message, Exception inner) : base(message, inner) { }
-}
+    #region Properties
+    // taken from the Linux programmer manual:
+    /// <summary>
+    /// Value that should be returned on a successful operation.
+    /// </summary>
+    private const int EXIT_SUCCESS = 0;
 
+    /// <summary>
+    /// Value that should be returned on a failed operation.
+    /// </summary>
+    private const int EXIT_FAILURE = 1;
 
+    /// <summary>
+    /// Value that determines if the current Program is running in interactive mode.
+    /// </summary>
+    private bool IsInteractive { get; }
 
-namespace UndertaleModCli
-{
+    /// <summary>
+    /// Value that determines if the current Program is running in verbose mode.
+    /// </summary>
+    private bool Verbose { get; }
 
-    public class LoadOptions
+    /// <summary>
+    /// File path or directory path that determines an output for the current Program.
+    /// </summary>
+    private FileSystemInfo Output { get; }
+
+    /// <summary>
+    /// Constant, used to indicate that the user wants to replace everything in a replace command.
+    /// </summary>
+    private const string UMT_REPLACE_ALL = "UMT_REPLACE_ALL";
+
+    /// <summary>
+    /// Constant, used to indicate that the user wants to dump everything in a dump command
+    /// </summary>
+    private const string UMT_DUMP_ALL = "UMT_DUMP_ALL";
+
+    //TODO: document these, these are intertwined with inherited updating methods
+    private int progressValue;
+    private Task updater;
+    private CancellationTokenSource cTokenSource;
+    private CancellationToken cToken;
+
+    private string savedMsg, savedStatus;
+    private double savedValue, savedValueMax;
+
+    /// <summary>
+    /// The ScriptOptions, only used for <see cref="CSharpScript"/>, aka running C# code.
+    /// </summary>
+    private ScriptOptions CliScriptOptions { get; }
+
+    /// <summary>
+    /// Determines if actions should show a "this is finished" text. Gets set by <see cref="SetFinishedMessage"/>.
+    /// </summary>
+    private bool FinishedMessageEnabled { get; set; }
+    #endregion
+
+    /// <summary>
+    /// Main entrypoint for Cli
+    /// </summary>
+    /// <param name="args">Arguments passed on to program.</param>
+    /// <returns>Result code of the program.</returns>
+    public static int Main(string[] args)
     {
+        var verboseOption = new Option<bool>(new []{"-v", "--verbose"}, "Detailed logs");
 
-        public FileInfo Datafile { get; set; }
-        public FileInfo[] Scripts { get; set; }
-        public string? Line { get; set; }
+        var dataFileArgument = new Argument<FileInfo>("datafile", "Path to the data.win/.ios/.droid/.unx file");
 
+        // Setup new command
+        Command newCommand = new Command("new", "Generates a blank data file")
+        {
+            new Option<FileInfo>(new []{"-o", "--output"}, () => new NewOptions().Output),
+            new Option<bool>(new []{"-f", "--overwrite"}, "Overwrite destination file if it already exists"),
+            new Option<bool>(new []{"-", "--stdout"}, "Write new data content to stdout"),  // "-" is often used in *nix land as a replacement for stdout
+            verboseOption
+        };
+        newCommand.Handler = CommandHandler.Create<NewOptions>(Program.New);
 
-        public FileInfo? Output { get; set; }
-        public bool Interactive { get; set; } = false;
-        public bool Verbose { get; set; } = false;
+        // Setup load command
+        var scriptRunnerOption = new Option<FileInfo[]>(new []{ "-s", "--scripts"}, "Scripts to apply to the <datafile>. Ex. a.csx b.csx");
+        Command loadCommand = new Command("load", "Load a data file and perform actions on it")
+        {
+            dataFileArgument,
+            scriptRunnerOption,
+            verboseOption,
+            //TODO: why no force overwrite here, but needed for new?
+            new Option<FileInfo>(new []{"-o", "--output"}, "Where to save the modified data file"),
+            new Option<string>(new []{"-l","--line"}, "Run C# string. Runs AFTER everything else"),
+            //TODO: make interactive another Command
+            new Option<bool>(new []{"-i", "--interactive"}, "Interactive menu launch")
+        };
+        loadCommand.Handler = CommandHandler.Create<LoadOptions>(Program.Load);
+
+        // Setup info command
+        Command infoCommand = new Command("info", "Show basic info about the game data file")
+        {
+            dataFileArgument,
+            verboseOption
+        };
+        infoCommand.Handler = CommandHandler.Create<InfoOptions>(Program.Info);
+
+        // Setup dump command
+        Command dumpCommand = new Command("dump", "Dump certain properties about the game data file")
+        {
+            dataFileArgument,
+            verboseOption,
+            new Option<DirectoryInfo>(new []{"-o", "--output"}, "Where to dump data file properties to. Will default to path of the data file"),
+            new Option<string[]>(new[] {"-c", "--code"},
+                $"The code files to dump. Ex. gml_Script_init_map gml_Script_reset_map. Specify '{UMT_DUMP_ALL}' to dump all code entries"),
+            new Option<bool>(new[] {"-s", "--strings"}, "Whether to dump all strings"),
+            new Option<bool>(new[] {"-t", "--textures"}, "Whether to dump all embedded textures")
+        };
+        dumpCommand.Handler = CommandHandler.Create<DumpOptions>(Program.Dump);
+
+        // Setup replace command
+        Command replaceCommand = new Command("replace", "Replace certain properties in the game data file")
+        {
+            dataFileArgument,
+            verboseOption,
+            new Option<FileInfo>(new []{"-o", "--output"}, "Where to save the modified data file"),
+            new Option<string[]>(new[] {"-c", "--code"},
+                $"Which code files to replace with which file. Ex. 'gml_Script_init_map=./newCode.gml'. It is possible to replace everything by using '{UMT_REPLACE_ALL}'"),
+            new Option<string[]>(new []{"-t", "--textures"},
+                $"Which embedded texture entry to replace with which file. Ex. 'Texture 0=./newTexture.png'. It is possible to replace everything by using '{UMT_REPLACE_ALL}'")
+        };
+        replaceCommand.Handler = CommandHandler.Create<ReplaceOptions>(Program.Replace);
+
+        // Merge everything together
+        RootCommand rootCommand = new RootCommand
+        {
+            newCommand,
+            loadCommand,
+            infoCommand,
+            dumpCommand,
+            replaceCommand
+        };
+        rootCommand.Description = "CLI tool for modding, decompiling and unpacking Undertale (and other Game Maker: Studio games)!";
+        Parser commandLine = new CommandLineBuilder(rootCommand)
+            .UseDefaults() // automatically configures dotnet-suggest
+            .Build();
+
+        return commandLine.Invoke(args);
     }
 
-    public class InfoOptions
+    public Program(FileInfo datafile, FileInfo[] scripts, FileInfo output, bool verbose = false, bool interactive = false)
     {
-        public FileInfo Datafile { get; set; }
-        public bool Verbose { get; set; } = false;
+        this.Verbose = verbose;
+        IsInteractive = interactive;
+        Console.InputEncoding = Encoding.UTF8;
+        Console.OutputEncoding = Console.InputEncoding;
+
+
+        Console.WriteLine($"Trying to load file: '{datafile.FullName}'");
+
+        this.FilePath = datafile.FullName;
+        this.ExePath = Environment.CurrentDirectory;
+        this.Output = output;
+
+        this.Data = ReadDataFile(datafile, this.Verbose ? WarningHandler : null, this.Verbose ? MessageHandler : null);
+
+        FinishedMessageEnabled = true;
+        this.CliScriptOptions = ScriptOptions.Default
+            .AddImports("UndertaleModLib", "UndertaleModLib.Models", "UndertaleModLib.Decompiler",
+                "UndertaleModLib.Scripting", "UndertaleModLib.Compiler",
+                "UndertaleModLib.Util", "System", "System.IO", "System.Collections.Generic",
+                "System.Text.RegularExpressions")
+            .AddReferences(typeof(UndertaleObject).GetTypeInfo().Assembly,
+                GetType().GetTypeInfo().Assembly,
+                typeof(JsonConvert).GetTypeInfo().Assembly,
+                typeof(System.Text.RegularExpressions.Regex).GetTypeInfo().Assembly,
+                typeof(TextureWorker).GetTypeInfo().Assembly)
+            // "WithEmitDebugInformation(true)" not only lets us to see a script line number which threw an exception,
+            // but also provides other useful debug info when we run UMT in "Debug".
+            .WithEmitDebugInformation(true);
+    }
+
+    public Program(FileInfo datafile, bool verbose, DirectoryInfo output = null)
+    {
+        if (datafile == null) throw new ArgumentNullException(nameof(datafile));
+
+        Console.WriteLine($"Trying to load file: '{datafile.FullName}'");
+        this.Verbose = verbose;
+        this.Data = ReadDataFile(datafile, verbose ? WarningHandler : null, verbose ? MessageHandler : null);
+        this.Output = output ?? new DirectoryInfo(datafile.DirectoryName);
+
+        if (this.Verbose)
+            Console.WriteLine("Output directory has been set to " + this.Output.FullName);
     }
 
     /// <summary>
-    /// Cli options for the New command
+    /// Method that gets executed on the "new" command
     /// </summary>
-    /// <param name="Overwrite">Save over an existing file at Output</param>
-    public class NewOptions
+    /// <param name="options">The arguments that have been provided with the "new" command</param>
+    /// <returns><see cref="EXIT_SUCCESS"/> and <see cref="EXIT_FAILURE"/> for being successful and failing respectively</returns>
+    private static int New(NewOptions options)
     {
-        /// <summary>
-        /// Destination for new data file
-        /// </summary>
-        public FileInfo Output { get; set; } = new FileInfo("data.win");
-        /// <summary>
-        /// Save over an existing file at Output
-        /// </summary>
-        public bool Overwrite { get; set; } = false;
-        /// <summary>
-        /// Whether to write the new data to Stdout
-        /// </summary>
-        public bool Stdout { get; set; }
+        //TODO: this should probably create a new Program instance, with just the properties that it needs
+
+        UndertaleData data = UndertaleData.CreateNew();
+
+        // If stdout flag is set, write new data to stdout and quit
+        if (options.Stdout)
+        {
+            if (options.Verbose) Console.WriteLine("Attempting to write new Data file to STDOUT...");
+            using MemoryStream ms = new MemoryStream();
+            UndertaleIO.Write(ms, data);
+            Console.OpenStandardOutput().Write(ms.ToArray(), 0, (int)ms.Length);
+            Console.Out.Flush();
+            if (options.Verbose) Console.WriteLine("Successfully wrote new Data file to STDOUT.");
+
+            return EXIT_SUCCESS;
+        }
+
+        // If not STDOUT, write to file instead. Check first if we have permission to overwrite
+        if (options.Output.Exists && !options.Overwrite)
+        {
+            Console.Error.WriteLine($"'{options.Output}' already exists. Pass --overwrite to overwrite");
+            return EXIT_FAILURE;
+        }
+
+        // We're not writing to STDOUT, and overwrite flag was given, so we write to specified file.
+        if (options.Verbose) Console.WriteLine($"Attempting to write new Data file to '{options.Output}'...");
+        using FileStream fs = options.Output.OpenWrite();
+        UndertaleIO.Write(fs, data);
+        if (options.Verbose) Console.WriteLine($"Successfully wrote new Data file to '{options.Output}'.");
+        return EXIT_SUCCESS;
     }
 
-    public class Program : IScriptInterface
+    /// <summary>
+    /// Method that gets executed on the "load" command
+    /// </summary>
+    /// <param name="options">The arguments that have been provided with the "load" command</param>
+    /// <returns><see cref="EXIT_SUCCESS"/> and <see cref="EXIT_FAILURE"/> for being successful and failing respectively</returns>
+    private static int Load(LoadOptions options)
     {
-        // taken fron the Linux programmer manual:
-        const int EXIT_SUCCESS = 0;
-        const int EXIT_FAILURE = 1;
-        public bool Interactive = false;
+        Program program;
 
-        private int progressValue;
-        private Task updater;
-        private CancellationTokenSource cts;
-        private CancellationToken cToken;
-
-        private string savedMsg, savedStatus;
-        private double savedValue, savedValueMax;
-
-        public bool GMLCacheEnabled => false; //not implemented yet
-
-        public bool IsAppClosed { get; set; }
-
-        public FileInfo? Output { get; set; }
-
-        /// <summary>
-        /// Read supplied filename and return the data file
-        /// </summary>
-        /// <param name="datafile"></param>
-        /// <param name="OnWarning"></param>
-        /// <param name="OnMessage"></param>
-        /// <returns></returns>
-        /// <exception cref="DataFileNotFoundException">If the data file cannot be found</exception>
-        public static UndertaleData ReadDataFile(FileInfo datafile, WarningHandlerDelegate? OnWarning = null, MessageHandlerDelegate? OnMessage = null)
+        // Try to load necessary values.
+        // This can throw if mandatory arguments are not given, in which case we want to exit cleanly without a stacktrace.
+        try
         {
-            try
-            {
-                using (var fs = datafile.OpenRead())
-                {
-                    return UndertaleIO.Read(fs, OnWarning, OnMessage);
-                }
-            }
-            catch (FileNotFoundException e)
-            {
-                throw new DataFileNotFoundException($"data file {e.FileName} does not exist");
-            }
-
+            program = new Program(options.Datafile, options.Scripts, options.Output, options.Verbose, options.Interactive);
         }
-        public bool Verbose { get; set; }
-
-        public Program(FileInfo datafile, FileInfo[]? scripts, FileInfo? output, bool verbose = false, bool interactive = false)
+        catch (Exception e)
         {
-            Verbose = verbose;
-            Interactive = interactive;
-            Console.InputEncoding = System.Text.Encoding.UTF8;
-            Console.OutputEncoding = Console.InputEncoding;
-
-
-            Console.WriteLine($"Trying to load file: {datafile.FullName}");
-
-
-
-            this.FilePath = datafile.FullName;
-            this.ExePath = Environment.CurrentDirectory;
-            this.Output = output;
-
-            this.Data = ReadDataFile(datafile, Verbose ? OnWarning : null, Verbose ? OnMessage : null);
-
-            FinishedMessageEnabled = true;
-            this.CliScriptOptions = ScriptOptions.Default
-                            .AddImports("UndertaleModLib", "UndertaleModLib.Models", "UndertaleModLib.Decompiler",
-                                        "UndertaleModLib.Scripting", "UndertaleModLib.Compiler",
-                                        "UndertaleModLib.Util", "System", "System.IO", "System.Collections.Generic",
-                                        "System.Text.RegularExpressions")
-                            .AddReferences(typeof(UndertaleObject).GetTypeInfo().Assembly,
-                                            GetType().GetTypeInfo().Assembly,
-                                            typeof(JsonConvert).GetTypeInfo().Assembly,
-                                            typeof(System.Text.RegularExpressions.Regex).GetTypeInfo().Assembly,
-                                            typeof(TextureWorker).GetTypeInfo().Assembly)
-                            // https://discord.com/channels/566861759210586112/568950566122946580/900145134480531506
-                            // ...WithEmitDebugInformation(true)" not only lets us to see a script line number which threw an exception,
-                            // but also provides other useful debug info when we run UMT in "Debug".
-                            .WithEmitDebugInformation(true);
+            Console.Error.WriteLine(e.Message);
+            return EXIT_FAILURE;
         }
 
-        public Program(FileInfo datafile, bool verbose)
+        // if interactive is enabled, launch the menu instead
+        if (options.Interactive)
         {
-            if (verbose)
-            {
-                this.Data = ReadDataFile(datafile, OnWarning, OnMessage);
-            }
+            program.RunInteractiveMenu();
+            return EXIT_SUCCESS;
+        }
+
+        // if we have any scripts to run, run every one of them
+        if (options.Scripts != null)
+        {
+            foreach (FileInfo script in options.Scripts)
+                program.RunCSharpFile(script.FullName);
+        }
+
+        // if line to execute was given, execute it
+        if (options.Line != null)
+        {
+            program.ScriptPath = null;
+            program.RunCSharpCode(options.Line);
+        }
+
+        // if parameter to save file was given, save the data file
+        if (options.Output != null)
+            program.SaveDataFile(options.Output.FullName);
+
+        return EXIT_SUCCESS;
+    }
+
+    /// <summary>
+    /// Method that gets executed on the "info" command
+    /// </summary>
+    /// <param name="options">The arguments that have been provided with the "info" command</param>
+    /// <returns><see cref="EXIT_SUCCESS"/> and <see cref="EXIT_FAILURE"/> for being successful and failing respectively</returns>
+    private static int Info(InfoOptions options)
+    {
+        Program program;
+        try
+        {
+            program = new Program(options.Datafile, options.Verbose);
+        }
+        catch (FileNotFoundException e)
+        {
+            Console.Error.WriteLine(e.Message);
+            return EXIT_FAILURE;
+        }
+
+        program.CliQuickInfo();
+        return EXIT_SUCCESS;
+    }
+
+    /// <summary>
+    /// Method that gets executed on the "dump" command
+    /// </summary>
+    /// <param name="options">The arguments that have been provided with the "dump" command</param>
+    /// <returns><see cref="EXIT_SUCCESS"/> and <see cref="EXIT_FAILURE"/> for being successful and failing respectively</returns>
+    private static int Dump(DumpOptions options)
+    {
+        Program program;
+        try
+        {
+            program = new Program(options.Datafile, options.Verbose, options.Output);
+        }
+        catch (FileNotFoundException e)
+        {
+            Console.Error.WriteLine(e.Message);
+            return EXIT_FAILURE;
+        }
+
+        // If user provided code to dump, dump code
+        if ((options.Code?.Length > 0) && (program.Data.Code.Count > 0))
+        {
+            // If user wanted to dump everything, do that, otherwise only dump what user provided
+            string[] codeArray;
+            if (options.Code.Contains(UMT_DUMP_ALL))
+                codeArray = program.Data.Code.Select(c => c.Name.Content).ToArray();
             else
-            {
-                this.Data = ReadDataFile(datafile, null, null);
-            }
+                codeArray = options.Code;
+
+            foreach (string code in codeArray)
+                program.DumpCodeEntry(code);
         }
 
-        public static int Main(string[] args)
+        // If user wanted to dump strings, dump all of them in a text file
+        if (options.Strings)
+            program.DumpAllStrings();
+
+        // If user wanted to dump embedded textures, dump all of them
+        if (options.Textures)
+            program.DumpAllTextures();
+
+        return EXIT_SUCCESS;
+    }
+
+    /// <summary>
+    /// Method that gets executed on the "replace" command
+    /// </summary>
+    /// <param name="options">The arguments that have been provided with the "replace" command</param>
+    /// <returns><see cref="EXIT_SUCCESS"/> and <see cref="EXIT_FAILURE"/> for being successful and failing respectively</returns>
+    private static int Replace(ReplaceOptions options)
+    {
+        Program program;
+        try
         {
-            var verboseOption = new Option<bool>(new []{"-v", "--verbose"}, "Detailed logs");
-
-            var dataFileOption = new Argument<FileInfo>("datafile")
-            {
-                Description = "Path to the data.win/.ios/.droid/.unx file"
-            };
-
-            var infoCommand = new Command("info", "Show info about game data file")
-            {
-                dataFileOption,
-                verboseOption,
-            };
-            infoCommand.Handler = CommandHandler.Create<InfoOptions>(Program.Info);
-
-            var scriptRunnerOption = new Option<FileInfo[]>(new []{ "-s", "--scripts"}, "Scripts to apply to the <datafile>. ex. a.csx b.csx");
-            var loadCommand = new Command("load", "Load data file and perform actions on it") {
-                dataFileOption,
-                scriptRunnerOption,
-                verboseOption,
-                new Option<FileInfo>(new []{"-o", "--output"}, "Where to save the modified data file"),
-                new Option<string>(new []{"-l","--line"}, "Run C# string. Runs AFTER everything else"),
-                new Option<bool>(new []{"-i", "--interactive"}, "Interactive menu launch"),
-
-            };
-            loadCommand.Handler = CommandHandler.Create<LoadOptions>(Program.Load);
-
-            var newCommand = new Command("new", "Generates a blank data file")
-            {
-                new Option<FileInfo>(new []{"-o", "--output"},getDefaultValue: () => new NewOptions().Output),
-                new Option<bool>(new []{"-f", "--overwrite"}, "Overwrite destination file if it already exists"),
-                new Option<bool>(new []{"-", "--stdout"}, "Write new data content to stdout"),  // "-" is often used in *nix land as a replacement for stdout
-            };
-            newCommand.Handler = CommandHandler.Create<NewOptions>(Program.New);
-
-            var rootCommand = new RootCommand {
-                infoCommand,
-                loadCommand,
-                newCommand,
-                };
-
-            rootCommand.Description = "CLI tool for modding, decompiling and unpacking Undertale (and other Game Maker: Studio games)!";
-            var commandLine = new CommandLineBuilder(rootCommand)
-                                    .UseDefaults() // automatically configures dotnet-suggest
-                                    .Build();
-
-            return commandLine.Invoke(args);
-
+            program = new Program(options.Datafile, null, options.Output, options.Verbose);
+        }
+        catch (FileNotFoundException e)
+        {
+            Console.Error.WriteLine(e.Message);
+            return EXIT_FAILURE;
         }
 
-        public static int New(NewOptions options)
+        // If user provided code to replace, replace them
+        if ((options.Code?.Length > 0) && (program.Data.Code.Count > 0))
         {
-            var data = UndertaleData.CreateNew();
-            if (options.Stdout)
+            // get the values and put them into a dictionary for ease of use
+            Dictionary<string, FileInfo> codeDict = new Dictionary<string, FileInfo>();
+            foreach (string code in options.Code)
             {
-                WriteStdout();
-            }
-            else
-            {
-                if (WriteFile() == EXIT_FAILURE)
+                string[] splitText = code.Split('=');
+
+                if (splitText.Length != 2)
                 {
+                    Console.Error.WriteLine($"{code} is malformed! Should be of format 'name_of_code=./newCode.gml' instead!");
                     return EXIT_FAILURE;
                 }
+
+                codeDict.Add(splitText[0], new FileInfo(splitText[1]));
             }
-            return EXIT_SUCCESS;
 
-
-            int WriteFile()
+            // If user wants to replace all, we'll be handling it differently. Replace every file from the provided directory
+            if (codeDict.ContainsKey(UMT_REPLACE_ALL))
             {
-                if (options.Output.Exists && !options.Overwrite)
+                string directory = codeDict[UMT_REPLACE_ALL].FullName;
+                foreach (FileInfo file in new DirectoryInfo(directory).GetFiles())
+                    program.ReplaceCodeEntryWithFile(file.Name, file);
+            }
+            // Otherwise, just replace every file which was given
+            else
+            {
+                foreach (KeyValuePair<string, FileInfo> keyValue in codeDict)
+                    program.ReplaceCodeEntryWithFile(keyValue.Key, keyValue.Value);
+            }
+        }
+
+        // If user provided texture to replace, replace them
+        if (options.Textures?.Length > 0)
+        {
+            // get the values and put them into a dictionary for ease of use
+            Dictionary<string, FileInfo> textureDict = new Dictionary<string, FileInfo>();
+            foreach (string texture in options.Textures)
+            {
+                string[] splitText = texture.Split('=');
+
+                if (splitText.Length != 2)
                 {
-                    Console.Error.WriteLine($"{options.Output} already exists. Pass --overwrite to overwrite");
+                    Console.Error.WriteLine($"{texture} is malformed! Should be of format 'Name=./new.png' instead!");
                     return EXIT_FAILURE;
                 }
-                using (var fs = options.Output.OpenWrite())
-                {
-                    UndertaleIO.Write(fs, data);
-                    return EXIT_SUCCESS;
-                }
+
+                textureDict.Add(splitText[0], new FileInfo(splitText[1]));
             }
 
-            void WriteStdout()
+            // If user wants to replace all, we'll be handling it differently. Replace every file from the provided directory
+            if (textureDict.ContainsKey(UMT_REPLACE_ALL))
             {
-                using (var ms = new MemoryStream())
-                {
-                    UndertaleIO.Write(ms, data);
-                    System.Console.OpenStandardOutput().Write(ms.ToArray(), 0, (int)ms.Length);
-                    System.Console.Out.Flush();
-                }
-
+                string directory = textureDict[UMT_REPLACE_ALL].FullName;
+                foreach (FileInfo file in new DirectoryInfo(directory).GetFiles())
+                    program.ReplaceTextureWithFile(file.Name, file);
+            }
+            // Otherwise, just replace every file which was given
+            else
+            {
+                foreach ((string key, FileInfo value) in textureDict)
+                    program.ReplaceTextureWithFile(key, value);
             }
         }
 
-        static public int Load(LoadOptions options)
+        // if parameter to save file was given, save the data file
+        if (options.Output != null)
+            program.SaveDataFile(options.Output.FullName);
+
+        return EXIT_SUCCESS;
+    }
+
+    /// <summary>
+    /// Runs the interactive menu indefinitely until user quits out of it.
+    /// </summary>
+    private void RunInteractiveMenu()
+    {
+        while (true)
         {
-
-            Program program;
-            try
-            {
-                program = new Program(options.Datafile, options.Scripts, options.Output, options.Verbose, options.Interactive);
-            }
-            catch (DataFileNotFoundException e)
-            {
-                Console.Error.WriteLine(e.Message);
-                return EXIT_FAILURE;
-            }
-            if (options.Interactive)
-            {
-                program.interactiveMenu();
-                return EXIT_SUCCESS;
-            }
-
-            if (options.Scripts != null)
-            {
-                foreach (var script in options.Scripts)
-                {
-                    program.RunCodeFile(script.FullName);
-                }
-            }
-
-            if (options.Line != null)
-            {
-                program.ScriptPath = null;
-                program.RunCodeLine(options.Line);
-            }
-
-            if (options.Output != null)
-            {
-                program.CliSave(options.Output.FullName);
-            }
-
-            return EXIT_SUCCESS;
-
-        }
-
-
-        public void interactiveMenu()
-        {
-            while (OnMenu()) { }
-        }
-
-        public UndertaleData Data { get; set; }
-
-        public string FilePath { get; set; }
-
-        public string ScriptPath { get; set; }
-
-        public object Highlighted { get; set; }
-
-        public object Selected { get; set; }
-
-        public bool CanSave { get; set; }
-
-        public bool ScriptExecutionSuccess { get; set; }
-
-        public string ScriptErrorMessage { get; set; }
-
-        public string ExePath { get; set; }
-
-        public string ScriptErrorType { get; set; }
-
-        public ScriptOptions CliScriptOptions { get; set; }
-
-        public bool FinishedMessageEnabled { get; set; }
-
-        // need this on Windows when drag and dropping files.
-        public static string Dequote(string a) => a.TrimStart('"').TrimEnd('"');
-
-        public void Pause()
-        {
-            // replica of the cmd pause command.
-            Console.Write("Press any key to continue . . . ");
-            Console.ReadKey(true);
-            Console.WriteLine();
-        }
-
-        public void RunCodeLine(string line)
-        {
-
-            try
-            {
-                CSharpScript.EvaluateAsync(line, CliScriptOptions, this, typeof(IScriptInterface)).GetAwaiter().GetResult();
-                ScriptExecutionSuccess = true;
-                ScriptErrorMessage = "";
-            }
-            catch (Exception exc)
-            {
-                ScriptExecutionSuccess = false;
-                ScriptErrorMessage = exc.ToString();
-                ScriptErrorType = "Exception";
-            }
-
-            if (FinishedMessageEnabled)
-            {
-                if (ScriptExecutionSuccess)
-                {
-                    string msg = $"Finished executing {ScriptPath ?? "C# line"}";
-                    if (Verbose)
-                        Console.WriteLine(msg);
-                }
-                else
-                {
-                    Console.Error.WriteLine(ScriptErrorMessage);
-                }
-            }
-        }
-
-        public void RunCodeFile(string path)
-        {
-            string lines;
-            try
-            {
-                lines = File.ReadAllText(path);
-            }
-            catch (IOException exc)
-            {
-                Console.Error.WriteLine("Script file not found or cannot be read.");
-                Console.Error.WriteLine(exc);
-                return;
-            }
-
-            if (lines != null)
-            {
-                ScriptPath = path;
-                RunCodeLine(lines);
-                ScriptPath = null;
-            }
-        }
-
-        public void OnWarning(string warning) => Console.WriteLine($"[WARNING]: {warning}");
-
-        public void OnMessage(string message) => Console.WriteLine($"[MESSAGE]: {message}");
-
-        public void CliSave(string to)
-        {
-            if (Verbose)
-            {
-                Console.WriteLine($"Saving new data file to {this.Output.FullName}");
-            }
-
-            using (var fs = new FileInfo(to).OpenWrite())
-            {
-                UndertaleIO.Write(fs, Data, OnMessage);
-                if (Verbose)
-                {
-                    Console.WriteLine($"Saved data file to {this.Output.FullName}");
-                }
-            }
-        }
-
-        public static int Info(InfoOptions options)
-        {
-            Program program;
-            try
-            {
-                program = new Program(options.Datafile, options.Verbose);
-            }
-            catch (DataFileNotFoundException e)
-            {
-                Console.Error.WriteLine(e.Message);
-                return EXIT_FAILURE;
-            }
-            program.CliQuickInfo();
-
-            return EXIT_SUCCESS;
-        }
-
-        public void CliQuickInfo()
-        {
-            Console.WriteLine("Quick Information:");
-            Console.WriteLine("Project Name - {0}", Data.GeneralInfo.Name);
-            Console.WriteLine("Is GMS2 - {0}", Data.IsGameMaker2());
-            Console.WriteLine("Is YYC - {0}", Data.IsYYC());
-            Console.WriteLine("Bytecode version - {0}", Data.GeneralInfo.BytecodeVersion);
-            Console.WriteLine("Configuration name - {0}", Data.GeneralInfo.Config);
-        }
-
-        public bool OnMenu()
-        {
+            Console.WriteLine("Interactive Menu:");
             Console.WriteLine("1 - Run script.");
             Console.WriteLine("2 - Run C# string.");
             Console.WriteLine("3 - Save and overwrite.");
             Console.WriteLine("4 - Save to different place.");
             Console.WriteLine("5 - Display quick info.");
-            Console.WriteLine("6 - Quit. (please be sure you've saved everything!!)");
+            //TODO: add dumping and replacing options
+            Console.WriteLine("6 - Quit without saving.");
 
             Console.Write("Input, please: ");
-            var k = Console.ReadKey().Key;
+            ConsoleKey k = Console.ReadKey().Key;
             Console.WriteLine();
 
             switch (k)
             {
+                // 1 - run script
+                case ConsoleKey.NumPad1:
                 case ConsoleKey.D1:
+                {
+                    Console.Write("File path (you can drag and drop)? ");
+                    string path = RemoveQuotes(Console.ReadLine());
+                    Console.WriteLine("Trying to run script {0}", path);
+                    try
                     {
-                        Console.Write("File path (you can drag and drop)? ");
-                        string path = Dequote(Console.ReadLine());
-                        Console.WriteLine("Trying to run script {0}", path);
-                        RunCodeFile(path);
-                        break;
+                        RunCSharpFile(path);
+                    }
+                    catch (Exception)
+                    {
+                        // ignored
                     }
 
+                    break;
+                }
+
+                // 2 - run c# string
+                case ConsoleKey.NumPad2:
                 case ConsoleKey.D2:
-                    {
-                        Console.Write("C# code line? ");
-                        string line = Console.ReadLine();
-                        ScriptPath = null;
-                        RunCodeLine(line);
-                        break;
-                    }
+                {
+                    Console.Write("C# code line? ");
+                    string line = Console.ReadLine();
+                    ScriptPath = null;
+                    RunCSharpCode(line);
+                    break;
+                }
 
+                // Save and overwrite data file
+                case ConsoleKey.NumPad3:
                 case ConsoleKey.D3:
-                    {
-                        CliSave(FilePath);
-                        break;
-                    }
+                {
+                    SaveDataFile(FilePath);
+                    break;
+                }
 
+                // Save data file to different path
+                case ConsoleKey.NumPad4:
                 case ConsoleKey.D4:
-                    {
-                        Console.Write("Where to save? ");
-                        string path = Dequote(Console.ReadLine());
-                        CliSave(path);
-                        break;
-                    }
+                {
+                    Console.Write("Where to save? ");
+                    string path = RemoveQuotes(Console.ReadLine());
+                    SaveDataFile(path);
+                    break;
+                }
 
+                // Print out Quick Info
+                case ConsoleKey.NumPad5:
                 case ConsoleKey.D5:
-                    {
-                        CliQuickInfo();
-                        Pause();
-                        break;
-                    }
+                {
+                    CliQuickInfo();
+                    break;
+                }
 
+                // Quit
+                case ConsoleKey.NumPad6:
                 case ConsoleKey.D6:
-                    {
-                        Console.WriteLine("Are you SURE? You can press 'n' and save before the changes are gone forever!!");
-                        Console.WriteLine("(Y/N)? ");
-                        var yes = Console.ReadKey(false).Key == ConsoleKey.Y;
-                        Console.WriteLine();
-                        if (yes) return false;
-                        break;
-                    }
+                {
+                    Console.WriteLine("Are you SURE? You can press 'n' and save before the changes are gone forever!!!");
+                    Console.WriteLine("(Y/N)? ");
+                    bool isInputYes = Console.ReadKey(false).Key == ConsoleKey.Y;
+                    Console.WriteLine();
+                    if (isInputYes) return;
+                    break;
+                }
 
                 default:
-                    {
-                        Console.WriteLine("Unknown input. Try using the upper line of digits on your keyboard.");
-                        break;
-                    }
-            }
-
-            return true;
-        }
-
-
-        public void EnsureDataLoaded()
-        {
-            if (Data is null)
-                throw new Exception("No data file is loaded.");
-        }
-
-        public async Task<bool> Make_New_File()
-        {
-            await Task.Delay(1); //dummy await
-
-            Data = UndertaleData.CreateNew();
-            Console.WriteLine("New file created.");
-            return true;
-        }
-
-        public void ReplaceTempWithMain(bool ImAnExpertBTW = false)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void ReplaceMainWithTemp(bool ImAnExpertBTW = false)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void ReplaceTempWithCorrections(bool ImAnExpertBTW = false)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void ReplaceCorrectionsWithTemp(bool ImAnExpertBTW = false)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void UpdateCorrections(bool ImAnExpertBTW = false)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void ScriptMessage(string message)
-        {
-            Console.WriteLine(message);
-            if (Interactive) { Pause(); }
-        }
-
-        public void SetUMTConsoleText(string message)
-        {
-            Console.Title = message;
-        }
-
-        public bool ScriptQuestion(string message)
-        {
-            Console.WriteLine(message);
-            Console.Write("Input (Y/N)? ");
-            var yes = Console.ReadKey(false).Key == ConsoleKey.Y;
-            Console.WriteLine();
-            return yes;
-        }
-
-        public void ScriptError(string error, string title = "Error", bool SetConsoleText = true)
-        {
-            // no need to care about SetConsoleText if we're in CLI.........
-            Console.WriteLine("--------------------------------------------------");
-            Console.WriteLine("---------------------ERROR!-----------------------");
-            Console.WriteLine("--------------------------------------------------");
-            Console.WriteLine(title);
-            Console.WriteLine("--------------------------------------------------");
-            Console.WriteLine(error);
-            Console.WriteLine("--------------------------------------------------");
-            Console.WriteLine("---------------------ERROR!-----------------------");
-            Console.WriteLine("--------------------------------------------------");
-            if (Interactive) { Pause(); }
-        }
-
-        public void ScriptOpenURL(string url)
-        {
-            Process p;
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                p = Process.Start(new ProcessStartInfo(url) { UseShellExecute = true }); // Works ok on windows
-            }
-            else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux) || RuntimeInformation.IsOSPlatform(OSPlatform.FreeBSD))
-            {
-                p = Process.Start("xdg-open", url);  // Works ok on linux, should work on FreeBSD as it's very similar.
-            }
-            else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-            {
-                p = Process.Start("open", url); // Not tested
-            }
-            else
-            {
-                throw new InvalidOperationException("Unable to open the browser on this OS.");
-            }
-
-            if (p != null) p.Dispose();
-        }
-
-        public bool SendAUMIMessage(IpcMessage_t ipMessage, ref IpcReply_t outReply)
-        {
-            return false;
-        }
-
-        public bool RunUMTScript(string path)
-        {
-            try
-            {
-                RunCodeFile(path);
-                return true;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-
-        public bool LintUMTScript(string path)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void InitializeScriptDialog()
-        {
-            throw new NotImplementedException();
-        }
-
-        public void ReapplyProfileCode()
-        {
-            throw new NotImplementedException();
-        }
-
-        public void NukeProfileGML(string codeName)
-        {
-            throw new NotImplementedException();
-        }
-
-        public string GetDecompiledText(string codeName, GlobalDecompileContext context = null)
-        {
-            throw new NotImplementedException();
-        }
-        public string GetDecompiledText(UndertaleCode code, GlobalDecompileContext context = null)
-        {
-            throw new NotImplementedException();
-        }
-
-        public string GetDisassemblyText(string codeName)
-        {
-            throw new NotImplementedException();
-        }
-        public string GetDisassemblyText(UndertaleCode code)
-        {
-            throw new NotImplementedException();
-        }
-
-        public bool AreFilesIdentical(string File01, string File02)
-        {
-            try
-            {
-                using var fs1 = new FileStream(File01, FileMode.Open, FileAccess.Read, FileShare.Read);
-                using var fs2 = new FileStream(File01, FileMode.Open, FileAccess.Read, FileShare.Read);
-                if (fs1.Length != fs2.Length) return false; // different size.
-
-                while (true)
                 {
-                    int b1 = fs1.ReadByte();
-                    int b2 = fs2.ReadByte();
-                    if (b1 != b2) return false; // different contents.
-                    if (b1 == -1 || b2 == -1) break; // end of files.
+                    Console.WriteLine("Unknown input. Try using the upper line of digits on your keyboard.");
+                    break;
                 }
-
-                // identical
-                return true;
             }
-            catch
+        }
+    }
+
+    /// <summary>
+    /// Prints some basic info about the loaded data file.
+    /// </summary>
+    private void CliQuickInfo()
+    {
+        Console.WriteLine("Quick Information:");
+        Console.WriteLine("Project Name - {0}", Data.GeneralInfo.Name);
+        Console.WriteLine("Is GMS2 - {0}", Data.IsGameMaker2());
+        Console.WriteLine("Is YYC - {0}", Data.IsYYC());
+        Console.WriteLine("Bytecode version - {0}", Data.GeneralInfo.BytecodeVersion);
+        Console.WriteLine("Configuration name - {0}", Data.GeneralInfo.Config);
+
+        Console.WriteLine($"{Data.Sounds.Count} Sounds, {Data.Sprites.Count} Sprites, {Data.Backgrounds.Count} Backgrounds");
+        Console.WriteLine($"{Data.Paths.Count} Paths, {Data.Scripts.Count} Scripts, {Data.Shaders.Count} Shaders");
+        Console.WriteLine($"{Data.Fonts.Count} Fonts, {Data.Timelines.Count} Timelines, {Data.GameObjects.Count} Game Objects");
+        Console.WriteLine($"{Data.Rooms.Count} Rooms, {Data.Extensions.Count} Extensions, {Data.TexturePageItems.Count} Texture Page Items");
+        Console.WriteLine($"{Data.Code.Count} Code Entries, {Data.Variables.Count} Variables, {Data.Functions.Count} Functions");
+        Console.WriteLine($"{Data.CodeLocals.Count} Code locals, {Data.Strings.Count} Strings, {Data.EmbeddedTextures.Count} Embedded Textures");
+        Console.WriteLine($"{Data.EmbeddedAudio.Count} Embedded Audio");
+
+        if (IsInteractive) Pause();
+    }
+
+    /// <summary>
+    /// Dumps a code entry from a data file.
+    /// </summary>
+    /// <param name="codeEntry">The code entry that should get dumped</param>
+    private void DumpCodeEntry(string codeEntry)
+    {
+        UndertaleCode code = Data.Code.ByName(codeEntry);
+
+
+        if (code == null)
+        {
+            Console.Error.WriteLine($"Data file does not contain a code entry named {codeEntry}!");
+            return;
+        }
+
+        string directory = $"{Output.FullName}/CodeEntries/";
+
+        Directory.CreateDirectory(directory);
+
+        if (Verbose)
+            Console.WriteLine($"Dumping {codeEntry}");
+
+        File.WriteAllText($"{directory}/{codeEntry}.gml", GetDecompiledText(code));
+    }
+
+    /// <summary>
+    /// Dumps all strings in a data file.
+    /// </summary>
+    private void DumpAllStrings()
+    {
+        string directory = Output.FullName;
+
+        Directory.CreateDirectory(directory);
+
+        StringBuilder combinedText = new StringBuilder();
+        foreach (UndertaleString dataString in Data.Strings)
+        {
+            if (Verbose)
+                Console.WriteLine($"Added {dataString.Content}");
+            combinedText.Append($"{dataString.Content}\n");
+        }
+
+        if (Verbose)
+            Console.WriteLine("Writing all strings to disk");
+        File.WriteAllText($"{directory}/strings.txt", combinedText.ToString());
+    }
+
+    /// <summary>
+    /// Dumps all embedded textures in a data file.
+    /// </summary>
+    private void DumpAllTextures()
+    {
+        string directory = $"{Output.FullName}/EmbeddedTextures/";
+
+        Directory.CreateDirectory(directory);
+
+        foreach (UndertaleEmbeddedTexture texture in Data.EmbeddedTextures)
+        {
+            if (Verbose)
+                Console.WriteLine($"Dumping {texture.Name}");
+            File.WriteAllBytes($"{directory}/{texture.Name.Content}.png", texture.TextureData.TextureBlob);
+        }
+    }
+
+    /// <summary>
+    /// Replaces a code entry with text from another file.
+    /// </summary>
+    /// <param name="codeEntry">The code entry to replace</param>
+    /// <param name="fileToReplace">File path which should replace the code entry.</param>
+    private void ReplaceCodeEntryWithFile(string codeEntry, FileInfo fileToReplace)
+    {
+        UndertaleCode code = Data.Code.ByName(codeEntry);
+
+        if (code == null)
+        {
+            Console.Error.WriteLine($"Data file does not contain a code entry named {codeEntry}!");
+            return;
+        }
+
+        if (Verbose)
+            Console.WriteLine("Replacing " + codeEntry);
+
+        ImportGMLString(codeEntry, File.ReadAllText(fileToReplace.FullName));
+    }
+
+    /// <summary>
+    /// Replaces an embedded texture with contents from another file.
+    /// </summary>
+    /// <param name="textureEntry">Embedded texture to replace</param>
+    /// <param name="fileToReplace">File path which should replace the embedded texture.</param>
+    private void ReplaceTextureWithFile(string textureEntry, FileInfo fileToReplace)
+    {
+        UndertaleEmbeddedTexture texture = Data.EmbeddedTextures.ByName(textureEntry);
+
+        if (texture == null)
+        {
+            Console.Error.WriteLine($"Data file does not contain an embedded texture named {textureEntry}!");
+            return;
+        }
+
+        if (Verbose)
+            Console.WriteLine("Replacing " + textureEntry);
+
+        texture.TextureData.TextureBlob = File.ReadAllBytes(fileToReplace.FullName);
+    }
+
+    /// <summary>
+    /// Evaluates and executes the contents of a file as C# Code.
+    /// </summary>
+    /// <param name="path">Path to file which contents to interpret as C# code</param>
+    private void RunCSharpFile(string path)
+    {
+        string lines;
+        try
+        {
+            lines = File.ReadAllText(path);
+        }
+        catch (Exception exc)
+        {
+            // rethrow as otherwise this will get interpreted as success
+            Console.Error.WriteLine(exc.Message);
+            throw;
+        }
+
+        ScriptPath = path;
+        RunCSharpCode(lines, ScriptPath);
+    }
+
+    /// <summary>
+    /// Evaluates and executes given C# code.
+    /// </summary>
+    /// <param name="code">The C# string to execute</param>
+    /// <param name="scriptFile">The path to the script file where <see cref="code"/> was executed from.
+    /// Leave as null, if it wasn't executed from a script file</param>
+    private void RunCSharpCode(string code, string scriptFile = null)
+    {
+        if (Verbose)
+            Console.WriteLine($"Attempting to execute '1{scriptFile ?? code}'...");
+
+        try
+        {
+            CSharpScript.EvaluateAsync(code, CliScriptOptions, this, typeof(IScriptInterface)).GetAwaiter().GetResult();
+            ScriptExecutionSuccess = true;
+            ScriptErrorMessage = "";
+        }
+        catch (Exception exc)
+        {
+            ScriptExecutionSuccess = false;
+            ScriptErrorMessage = exc.ToString();
+            ScriptErrorType = "Exception";
+        }
+
+        if (!FinishedMessageEnabled) return;
+
+        if (ScriptExecutionSuccess)
+        {
+            if (Verbose)
+                Console.WriteLine($"Finished executing '{scriptFile ?? code}'");
+        }
+        else
+        {
+            Console.Error.WriteLine(ScriptErrorMessage);
+        }
+    }
+
+    /// <summary>
+    /// Saves the currently loaded <see cref="Data"/> to an output path.
+    /// </summary>
+    /// <param name="outputPath">The path where to save the data.</param>
+    private void SaveDataFile(string outputPath)
+    {
+        if (Verbose)
+            Console.WriteLine($"Saving new data file to '{outputPath}'");
+
+        using FileStream fs = new FileInfo(outputPath).OpenWrite();
+        UndertaleIO.Write(fs, Data, MessageHandler);
+        if (Verbose)
+            Console.WriteLine($"Saved data file to '{outputPath}'");
+    }
+
+    /// <summary>
+    /// Read supplied filename and return the data file.
+    /// </summary>
+    /// <param name="datafile">The datafile to read</param>
+    /// <param name="warningHandler">Handler for Warnings</param>
+    /// <param name="messageHandler">Handler for Messages</param>
+    /// <returns></returns>
+    /// <exception cref="FileNotFoundException">If the data file cannot be found</exception>
+    private static UndertaleData ReadDataFile(FileInfo datafile, WarningHandlerDelegate warningHandler = null, MessageHandlerDelegate messageHandler = null)
+    {
+        try
+        {
+            using FileStream fs = datafile.OpenRead();
+            UndertaleData gmData = UndertaleIO.Read(fs, warningHandler, messageHandler);
+            return gmData;
+        }
+        catch (FileNotFoundException e)
+        {
+            throw new FileNotFoundException($"Data file '{e.FileName}' does not exist");
+        }
+    }
+
+    // need this on Windows when drag and dropping files.
+    /// <summary>
+    /// Trims <c>"</c> or <c>'</c> from the beginning and end of a string.
+    /// </summary>
+    /// <param name="s"><see cref="String"/> to remove <c>"</c> and/or <c>'</c> from</param>
+    /// <returns>A new <see cref="String"/> that can be directly passed onto a FileInfo Constructor</returns>
+    //TODO: needs some proper testing on how it behaves on Linux/MacOS and might need to get expanded
+    private static string RemoveQuotes(string s)
+
+    {
+        return s.TrimStart('"').TrimEnd('"');
+    }
+
+    /// <summary>
+    /// Replicated the CMD Pause command. Waits for any key to be pressed before continuing.
+    /// </summary>
+    private static void Pause()
+    {
+        Console.WriteLine();
+        Console.Write("Press any key to continue . . . ");
+        Console.ReadKey(true);
+        Console.WriteLine();
+    }
+
+    /// <summary>
+    /// A simple warning handler that prints warnings to console.
+    /// </summary>
+    /// <param name="warning">The warning to print</param>
+    private static void WarningHandler(string warning) => Console.WriteLine($"[WARNING]: {warning}");
+
+    /// <summary>
+    /// A simple message handler that prints messages to console.
+    /// </summary>
+    /// <param name="message">The message to print</param>
+    private static void MessageHandler(string message) => Console.WriteLine($"[MESSAGE]: {message}");
+
+    //TODO: document these as well
+    private void ProgressUpdater()
+    {
+        DateTime prevTime = default;
+        int prevValue = 0;
+
+        while (true)
+        {
+            if (cToken.IsCancellationRequested)
             {
-                // wtf?!
-                return false;
+                if (prevValue >= progressValue) //if reached maximum
+                    return;
+
+                if (prevTime == default)
+                    prevTime = DateTime.UtcNow;                                       //begin measuring
+                else if (DateTime.UtcNow.Subtract(prevTime).TotalMilliseconds >= 500) //timeout - 0.5 seconds
+                    return;
             }
-        }
 
-        public string ScriptInputDialog(string titleText, string labelText, string defaultInputBoxText, string cancelButtonText, string submitButtonText, bool isMultiline, bool preventClose)
-        {
-            throw new NotImplementedException();
-        }
+            UpdateProgressValue(progressValue);
 
-        public string SimpleTextInput(string title, string label, string defaultValue, bool allowMultiline, bool showDialog = true)
-        {
-            throw new NotImplementedException();
-        }
-        public void SimpleTextOutput(string title, string label, string defaultText, bool allowMultiline)
-        {
-            throw new NotImplementedException();
-        }
-        public async Task ClickableTextOutput(string title, string query, int resultsCount, IOrderedEnumerable<KeyValuePair<string, List<string>>> resultsDict, bool editorDecompile, IOrderedEnumerable<string> failedList = null)
-        {
-            await Task.Delay(1); //dummy await
-            throw new NotImplementedException();
-        }
-        public async Task ClickableTextOutput(string title, string query, int resultsCount, IDictionary<string, List<string>> resultsDict, bool editorDecompile, IEnumerable<string> failedList = null)
-        {
-            await Task.Delay(1); //dummy await
-            throw new NotImplementedException();
-        }
+            prevValue = progressValue;
 
-        public void SetFinishedMessage(bool isFinishedMessageEnabled)
-        {
-            FinishedMessageEnabled = isFinishedMessageEnabled;
+            Thread.Sleep(100); //10 times per second
         }
-        public void UpdateProgressBar(string message, string status, double progressValue, double maxValue)
-        {
-            // i know, ugly
-            if (String.IsNullOrEmpty(message))
-                Console.WriteLine($"[{status}] {progressValue} out of {maxValue}");
-            else
-                Console.WriteLine($"[{message}|{status}] {progressValue} out of {maxValue}");
-        }
-
-        public void SetProgressBar(string message, string status, double progressValue, double maxValue)
-        {
-            savedMsg = message;
-            savedStatus = status;
-            savedValue = progressValue;
-            savedValueMax = maxValue;
-
-            UpdateProgressBar(message, status, progressValue, maxValue);
-        }
-        public void SetProgressBar()
-        {
-            //no dialog to show
-        }
-
-        public void UpdateProgressValue(double progressValue)
-        {
-            UpdateProgressBar(savedMsg, savedStatus, progressValue, savedValueMax);
-
-            savedValue = progressValue;
-        }
-        public void UpdateProgressStatus(string status)
-        {
-            UpdateProgressBar(savedMsg, status, savedValue, savedValueMax);
-
-            savedStatus = status;
-        }
-
-        public void HideProgressBar()
-        {
-            // nothing to hide..
-        }
-
-        public void AddProgress(int amount)
-        {
-            progressValue += amount;
-        }
-        public void IncProgress()
-        {
-            progressValue++;
-        }
-        public void AddProgressP(int amount) //P - Parallel (multithreaded)
-        {
-            Interlocked.Add(ref progressValue, amount); //thread-safe add operation (not the same as "lock ()")
-        }
-        public void IncProgressP()
-        {
-            Interlocked.Increment(ref progressValue); //thread-safe increment
-        }
-        public int GetProgress()
-        {
-            return progressValue;
-        }
-        public void SetProgress(int value)
-        {
-            progressValue = value;
-        }
-
-        public void EnableUI()
-        {
-            // nothing to enable...
-        }
-
-        public void SyncBinding(string resourceType, bool enable)
-        {
-            //there is no UI with any data binding
-        }
-        public void SyncBinding(bool enable = false)
-        {
-            //there is no UI with any data binding
-        }
-
-        private void ProgressUpdater()
-        {
-            DateTime prevTime = default;
-            int prevValue = 0;
-
-            while (true)
-            {
-                if (cToken.IsCancellationRequested)
-                {
-                    if (prevValue >= progressValue) //if reached maximum
-                        return;
-                    else
-                    {
-                        if (prevTime == default)
-                            prevTime = DateTime.UtcNow;                                       //begin measuring
-                        else if (DateTime.UtcNow.Subtract(prevTime).TotalMilliseconds >= 500) //timeout - 0.5 seconds
-                            return;
-                    }
-                }
-
-                UpdateProgressValue(progressValue);
-
-                prevValue = progressValue;
-
-                Thread.Sleep(100); //10 times per second
-            }
-        }
-        public void StartUpdater()
-        {
-            if (cts is not null)
-                Console.WriteLine("Warning - there is another progress updater task running (hangs) in the background.");
-
-            cts = new CancellationTokenSource();
-            cToken = cts.Token;
-
-            updater = Task.Run(ProgressUpdater);
-        }
-        public async Task StopUpdater() //"async" because "Wait()" blocks UI thread
-        {
-            if (cts is not null)
-            {
-                cts.Cancel();
-
-                if (await Task.Run(() => !updater.Wait(2000))) //if ProgressUpdater isn't responding
-                    Console.WriteLine("Error - stopping the progress updater task is failed.");
-                else
-                {
-                    cts.Dispose();
-                    cts = null;
-                }
-
-                updater.Dispose();
-            }
-        }
-
-        public async Task<bool> GenerateGMLCache(ThreadLocal<GlobalDecompileContext> decompileContext = null, object dialog = null, bool isSaving = false)
-        {
-            await Task.Delay(1); //dummy await
-
-            //not implemented yet
-
-            return false;
-        }
-
-
-        public void ChangeSelection(object newsel)
-        {
-            Selected = newsel;
-        }
-
-        public string PromptChooseDirectory(string prompt)
-        {
-            Console.WriteLine("Please type a path (or drag and drop) to a directory:");
-            Console.Write("Path: ");
-            string p = Console.ReadLine();
-            return p;
-        }
-
-        public string PromptLoadFile(string defaultExt, string filter)
-        {
-            throw new NotImplementedException("Sorry, this hasn't been implemented yet!");
-        }
-
-        public void ImportGMLString(string codeName, string gmlCode, bool doParse = true, bool CheckDecompiler = false)
-        {
-            throw new NotImplementedException("Sorry, this hasn't been implemented yet!");
-        }
-
-        public void ImportASMString(string codeName, string gmlCode, bool doParse = true, bool destroyASM = true, bool CheckDecompiler = false)
-        {
-            throw new NotImplementedException("Sorry, this hasn't been implemented yet!");
-        }
-
-        public void ImportGMLFile(string fileName, bool doParse = true, bool CheckDecompiler = false, bool throwOnError = false)
-        {
-            throw new NotImplementedException("Sorry, this hasn't been implemented yet!");
-        }
-
-        public void ImportASMFile(string fileName, bool doParse = true, bool destroyASM = true, bool CheckDecompiler = false, bool throwOnError = false)
-        {
-            throw new NotImplementedException("Sorry, this hasn't been implemented yet!");
-        }
-
-        public void ReplaceTextInGML(string codeName, string keyword, string replacement, bool case_sensitive = false, bool isRegex = false, GlobalDecompileContext context = null)
-        {
-            throw new NotImplementedException("Sorry, this hasn't been implemented yet!");
-        }
-        public void ReplaceTextInGML(UndertaleCode code, string keyword, string replacement, bool case_sensitive = false, bool isRegex = false, GlobalDecompileContext context = null)
-        {
-            throw new NotImplementedException("Sorry, this hasn't been implemented yet!");
-        }
-
-        public bool DummyBool()
-        {
-            return false;
-        }
-
-        public void DummyVoid()
-        {
-            // i want hugs so bad :(
-        }
-
-        public string DummyString()
-        {
-            return "";
-        }
-
     }
 }
