@@ -1,18 +1,15 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using UndertaleModLib.Models;
 
 namespace UndertaleModLib
 {
     public class UndertaleSimpleList<T> : ObservableCollection<T>, UndertaleObject where T : UndertaleObject, new()
     {
+        /// <inheritdoc />
         public void Serialize(UndertaleWriter writer)
         {
             writer.Write((uint)Count);
@@ -29,6 +26,7 @@ namespace UndertaleModLib
             }
         }
 
+        /// <inheritdoc />
         public void Unserialize(UndertaleReader reader)
         {
             uint count = reader.ReadUInt32();
@@ -49,6 +47,7 @@ namespace UndertaleModLib
 
     public class UndertaleSimpleListString : ObservableCollection<UndertaleString>, UndertaleObject
     {
+        /// <inheritdoc />
         public void Serialize(UndertaleWriter writer)
         {
             writer.Write((uint)Count);
@@ -65,6 +64,7 @@ namespace UndertaleModLib
             }
         }
 
+        /// <inheritdoc />
         public void Unserialize(UndertaleReader reader)
         {
             uint count = reader.ReadUInt32();
@@ -96,6 +96,7 @@ namespace UndertaleModLib
                 throw new InvalidOperationException("Count of short SimpleList exceeds maximum number allowed.");
         }
 
+        /// <inheritdoc />
         public void Serialize(UndertaleWriter writer)
         {
             writer.Write((ushort)Count);
@@ -112,6 +113,7 @@ namespace UndertaleModLib
             }
         }
 
+        /// <inheritdoc />
         public void Unserialize(UndertaleReader reader)
         {
             ushort count = reader.ReadUInt16();
@@ -132,6 +134,7 @@ namespace UndertaleModLib
 
     public class UndertalePointerList<T> : ObservableCollection<T>, UndertaleObject where T : UndertaleObject, new()
     {
+        /// <inheritdoc />
         public void Serialize(UndertaleWriter writer)
         {
             writer.Write((uint)Count);
@@ -167,6 +170,7 @@ namespace UndertaleModLib
             }
         }
 
+        /// <inheritdoc />
         public void Unserialize(UndertaleReader reader)
         {
             uint count = reader.ReadUInt32();
@@ -212,6 +216,7 @@ namespace UndertaleModLib
 
     public class UndertalePointerListLenCheck<T> : UndertalePointerList<T>, UndertaleObjectEndPos where T : UndertaleObjectLenCheck, new()
     {
+        /// <inheritdoc />
         public void Unserialize(UndertaleReader reader, uint endPosition)
         {
             uint count = reader.ReadUInt32();
