@@ -9,11 +9,11 @@ namespace UndertaleModTool
     //Make new GUID helper functions
     public partial class MainWindow : Window, INotifyPropertyChanged, IScriptInterface
     {
-        public void ReplaceTempWithMain(bool ImAnExpertBTW = false)
+        public void ReplaceTempWithMain(bool imAnExpertBtw = false)
         {
             try
             {
-                if (!ImAnExpertBTW && (!(ScriptQuestion("Warning: This may cause desyncs! The intended purpose is for reverting incorrect code corrections.\nOnly use this if you know what you're doing! Continue?"))))
+                if (!imAnExpertBtw && (!(ScriptQuestion("Warning: This may cause desyncs! The intended purpose is for reverting incorrect code corrections.\nOnly use this if you know what you're doing! Continue?"))))
                     return;
                 string MainPath = Path.Combine(ProfilesFolder, ProfileHash, "Main");
                 string TempPath = Path.Combine(ProfilesFolder, ProfileHash, "Temp");
@@ -28,11 +28,11 @@ namespace UndertaleModTool
                 MessageBox.Show("ReplaceTempWithMain error! Send this to Grossley#2869 and make an issue on Github\n" + exc.ToString());
             }
         }
-        public void ReplaceMainWithTemp(bool ImAnExpertBTW = false)
+        public void ReplaceMainWithTemp(bool imAnExpertBtw = false)
         {
             try
             {
-                if (!ImAnExpertBTW && (!(ScriptQuestion("Warning: This may cause desyncs! The intended purpose is for pushing code corrections (such as asset resolutions)\nOnly use this if you know what you're doing! Continue?"))))
+                if (!imAnExpertBtw && (!(ScriptQuestion("Warning: This may cause desyncs! The intended purpose is for pushing code corrections (such as asset resolutions)\nOnly use this if you know what you're doing! Continue?"))))
                     return;
                 string MainPath = Path.Combine(ProfilesFolder, ProfileHash, "Main");
                 string TempPath = Path.Combine(ProfilesFolder, ProfileHash, "Temp");
@@ -47,11 +47,11 @@ namespace UndertaleModTool
                 MessageBox.Show("ReplaceMainWithTemp error! Send this to Grossley#2869 and make an issue on Github\n" + exc.ToString());
             }
         }
-        public void ReplaceTempWithCorrections(bool ImAnExpertBTW = false)
+        public void ReplaceTempWithCorrections(bool imAnExpertBtw = false)
         {
             try
             {
-                if (!ImAnExpertBTW && (!(ScriptQuestion("If you messed up royally while developing your corrections, you can use this to revert all of the changes to them in your Temp folder to what is in the Corrections folder. Only use this if you know what you're doing! Continue?"))))
+                if (!imAnExpertBtw && (!(ScriptQuestion("If you messed up royally while developing your corrections, you can use this to revert all of the changes to them in your Temp folder to what is in the Corrections folder. Only use this if you know what you're doing! Continue?"))))
                     return;
                 string MainPath = Path.Combine(ProfilesFolder, ProfileHash, "Main");
                 string TempPath = Path.Combine(ProfilesFolder, ProfileHash, "Temp");
@@ -72,11 +72,11 @@ namespace UndertaleModTool
                 MessageBox.Show("ReplaceCorrectionsWithTemp error! Send this to Grossley#2869 and make an issue on Github\n" + exc.ToString());
             }
         }
-        public void ReplaceCorrectionsWithTemp(bool ImAnExpertBTW = false)
+        public void ReplaceCorrectionsWithTemp(bool imAnExpertBtw = false)
         {
             try
             {
-                if (!ImAnExpertBTW && (!(ScriptQuestion("The intended purpose is for pushing your custom made code corrections (your entire temporary profile) to the code corrections folder bundled with UndertaleModTool, making them permanent. Only use this if you know what you're doing! Continue?"))))
+                if (!imAnExpertBtw && (!(ScriptQuestion("The intended purpose is for pushing your custom made code corrections (your entire temporary profile) to the code corrections folder bundled with UndertaleModTool, making them permanent. Only use this if you know what you're doing! Continue?"))))
                     return;
                 Directory.Delete(CorrectionsFolder, true);
                 string MainPath = Path.Combine(ProfilesFolder, ProfileHash, "Main");
@@ -88,13 +88,13 @@ namespace UndertaleModTool
                 MessageBox.Show("ReplaceCorrectionsWithTemp error! Send this to Grossley#2869 and make an issue on Github\n" + exc.ToString());
             }
         }
-        public void UpdateCorrections(bool ImAnExpertBTW = false)
+        public void UpdateCorrections(bool imAnExpertBtw = false)
         {
-            if (!ImAnExpertBTW && (!(ScriptQuestion("Update Main with Temp, Temp with Main, Corrections with Temp. Only use this if you know what you're doing! Continue?"))))
+            if (!imAnExpertBtw && (!(ScriptQuestion("Update Main with Temp, Temp with Main, Corrections with Temp. Only use this if you know what you're doing! Continue?"))))
                 return;
-            ReplaceMainWithTemp(ImAnExpertBTW);
-            ReplaceTempWithMain(ImAnExpertBTW);
-            ReplaceCorrectionsWithTemp(ImAnExpertBTW);
+            ReplaceMainWithTemp(imAnExpertBtw);
+            ReplaceTempWithMain(imAnExpertBtw);
+            ReplaceCorrectionsWithTemp(imAnExpertBtw);
         }
     }
 }
